==> 1_data_preprocessing.Rmd <==
---
title: "1_data_preprocessing"
author: "Joy_Fu"
date: "2023-05-01"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
rm(list = ls())
pacman::p_load(tidyverse, GenomicRanges, rtracklayer, bigsnpr, bigreadr, 
               data.table, ieugwasr)
raw_data_path = "/Users/Mingzhou/Desktop/Projects/Dementia-prediction/data/"
output_path = "/Users/Mingzhou/Desktop/Projects/Dementia-prediction/output/"
# Source in useful functions
source("functions.R")
```

# Part 0. GWAS preprocessing
Preprocessing of GWAS summary statistics, ran once
## 1. AD Kunkle EUR
```{r}
sumstats_file = paste0(raw_data_path, 
                       "sumstats/raw/Kunkle2019_AD_EUR_Stage1.txt.gz")
sumstats = fread(sumstats_file)
names(sumstats) = c("chr", "pos", "rsid", "a1", "a0", "beta", "se", "p")
sumstats_clean = sumstats %>% filter(!is.na(chr) & !is.na(pos)) %>% 
  mutate(p = as.numeric(p))
dim(sumstats_clean) # dim = (11480632,8)
write.table(sumstats_clean, file = paste0(raw_data_path, 'sumstats/fuma_upload/AD_Kunkle2019_EUR_hg19.txt'),
            sep = "\t", quote = F, row.names = F, col.names = T)
```

## 2. AD Kunkle AFR
```{r}
sumstats_file = paste0(raw_data_path, 
                       "sumstats/raw/Kunkle2020_AD_AA_META_Model1.txt.gz")
sumstats = fread(sumstats_file)
names(sumstats) = c("chr", "pos", "marker", "a1", "a0", 
                    "beta", "se", "p", "a1_freq")
sumstats_clean = sumstats %>% filter(!is.na(chr) & !is.na(pos))
dim(sumstats_clean) # dim = (27724505,9)
write.table(sumstats_clean, file = paste0(raw_data_path, 'sumstats/fuma_upload/AD_Kunkle2021_AFR_hg19.txt'),
            sep = "\t", quote = F, row.names = F, col.names = T)
```

## 3. AD Jun Transethnic
```{r}
sumstats_file = paste0(raw_data_path, 
                       "sumstats/raw/Jun2017_AD_Transethnic_META_ALL.sample.tbx.gz")
sumstats = fread(sumstats_file)
names(sumstats) = c("chr", "pos", "marker", "a1", "a0", "a1_freq", "freqse", 
                    "beta", "se", "p", "direction", "hetchisq", "hetdf", "hetp")
sumstats_clean = sumstats %>% filter(!is.na(chr) & !is.na(pos)) %>% 
  mutate(a1 = toupper(a1), a0 = toupper(a0), p = as.numeric(p))
dim(sumstats_clean) # dim = (7343832,14)
write.table(sumstats_clean, 
            file = paste0(raw_data_path, 
                          'sumstats/fuma_upload/AD_Jun2017_Trans_hg19.txt'),
            sep = "\t", quote = F, row.names = F, col.names = T)
```

## 4. LBD
```{r}
sumstats_file = paste0(raw_data_path, "sumstats/raw/Chia2021_LBD_EUR.txt.gz")
sumstats = fread(sumstats_file) 
names(sumstats) = c("chr", "pos", "rsid", "a1", "a0", "beta", "se", "p")
sumstats_clean = sumstats %>% filter(!is.na(chr) & !is.na(pos)) 
dim(sumstats_clean) # dim = (7827366,8)
write.table(sumstats_clean, 
            file = paste0(raw_data_path, 
                          'sumstats/fuma_upload/LBD_Chia2021_hg19.txt'),
            sep = "\t", quote = F, row.names = F, col.names = T)
```

## 5. PD
```{r}
sumstats_file = paste0(raw_data_path, 
                       "sumstats/raw/Nalls2019_PD_EUR_excluding23andMe.tab.gz")
sumstats = fread(sumstats_file)
sumstats_add = sumstats %>% separate(col = "SNP", into = c("chr_pre", "pos"), 
                                     sep = ":", remove = F) %>% 
  mutate(chr = as.integer(substring(chr_pre, 4, last = 1000000L)), 
         pos = as.integer(pos)) %>% 
  mutate(N_total = N_cases + N_controls) %>% 
  select(chr, pos, SNP, A1, A2, freq, b, se, p, N_cases, N_controls, N_total)
names(sumstats_add) = c("chr", "pos", "marker", "a1", "a0", "a1_freq",
                        "beta", "se", "p", "n_case", "n_control", "n_total")
sumstats_clean = sumstats_add %>% filter(!is.na(pos) & !is.na(chr)) 
dim(sumstats_clean) # dim = (17510617,12)
write.table(sumstats_clean, 
            file = paste0(raw_data_path, 
                          'sumstats/fuma_upload/PD_Nalls2019_hg19.txt'),
            sep = "\t", quote = F, row.names = F, col.names = T)
```

## 6. PSP
```{bash}
# Need some preprocess of the file first in the bash
sed '/CN[0-9]/d' Chen2018_PSP_EUR.tab > Chen2018_PSP_EUR_rm.tab
sed '/INS:/d' Chen2018_PSP_EUR_rm.tab > Chen2018_PSP_EUR_rm2.tab
sed '/INV/d' Chen2018_PSP_EUR_rm2.tab > Chen2018_PSP_EUR_rm3.tab
```

```{r}
sumstats_file = paste0(raw_data_path, "sumstats/raw/Chen2018_PSP_EUR_rm3.tab")
sumstats = fread(sumstats_file)
names(sumstats)[1] = "marker"
names(sumstats)[3] = "pos"
sumstats_clean = sumstats %>% 
  mutate(rsid = case_when(
    nchar(SNP) >= 3 ~ SNP
  )) %>%  
  select(CHR, pos, rsid, ALLELE1, ALLELE0, A1FREQ, BETA, SE, P_BOLT_LMM_INF) %>% 
  filter(!is.na(CHR) & !is.na(pos)) 
names(sumstats_clean) = c("chr", "pos", "rsid", "a1", "a0", "a1_freq", 
                          "beta", "se", "p")
dim(sumstats_clean) # dim = (6417560,9)
write.table(sumstats_clean, 
            file = paste0(raw_data_path, 
                          'sumstats/fuma_upload/PSP_Chen2018_hg19.txt'),
            sep = "\t", quote = F, row.names = F, col.names = T)
```

## 7. Stroke
```{r}
sumstats_file = paste0(raw_data_path, "sumstats/raw/MEGASTROKE.1.AS.TRANS.out")
sumstats = fread(sumstats_file)
names(sumstats) = c("rsid", "a1", "a0", "a1_freq", "beta", "se", "p")
sumstats_clean = sumstats %>% filter(!is.na(rsid)) %>% 
  mutate(a1 = toupper(a1), a0 = toupper(a0))
dim(sumstats_clean) # dim = (7675830,7)
write.table(sumstats_clean, 
            file = paste0(raw_data_path, 
                          'sumstats/fuma_upload/Stroke_MEGA_hg19.txt'),
            sep = "\t", quote = F, row.names = F, col.names = T)
```

# Part 1. ATLAS SNPs candidate extraction
## 1. Extract all SNPs overlapped with 1000G
```{r}
# read in txt file
atlas_snps = fread(paste0(raw_data_path, "atlas/geno/atlas.snplist.txt"), header = F)
names(atlas_snps) = c("chr", "marker", "pos", "alt", "ref")
dim(atlas_snps) # dim = (10388380, 5)
```

## 2. Clean dataframe to upload to AOU (check overlaps)
```{r}
options(scipen = 999)
atlas_snps_sep = atlas_snps %>% mutate(chr = paste0("chr", chr)) %>%
  mutate(start = as.integer(pos)) %>% filter(!is.na(pos) & !is.na(chr)) %>% 
  mutate(stop = start + 1) %>% dplyr::rename(rsid = marker) %>% 
  select(chr, start, stop, rsid, ref, alt) %>% 
  mutate(start = as.character(start), stop = as.character(stop))
dim(atlas_snps_sep) # dim = (10388380, 6)
# output to txt (check overlaps on AOU and return SNPs that shared by both)
write.table(atlas_snps_sep, 
            file = paste0(raw_data_path, 'atlas/geno/atlas.intersectsnp.clean.txt'),
            sep = "\t", quote = F, row.names = F, col.names = T)
```

```{r}
# atlas_snps_sep_short = atlas_snps_sep[8310705:10388380,]
# write.table(atlas_snps_sep_short, 
#             file = paste0(raw_data_path, 'atlas/geno/atlas.intersectsnp.clean.short5.txt'),
#             sep = "\t", quote = F, row.names = F, col.names = T)
```

```{r}
# read in overlapped snps
atlas_snps_overlap1 = fread(paste0(raw_data_path, "atlas/geno/atlas_aou_overlap1.txt"))
atlas_snps_overlap2 = fread(paste0(raw_data_path, "atlas/geno/atlas_aou_overlap2.txt"))
atlas_snps_overlap3 = fread(paste0(raw_data_path, "atlas/geno/atlas_aou_overlap3.txt"))
atlas_snps_overlap4 = fread(paste0(raw_data_path, "atlas/geno/atlas_aou_overlap4.txt"))
atlas_snps_overlap5 = fread(paste0(raw_data_path, "atlas/geno/atlas_aou_overlap5.txt"))
atlas_aou_overlap_full = rbind(atlas_snps_overlap1, atlas_snps_overlap2,
                               atlas_snps_overlap3, atlas_snps_overlap4,
                               atlas_snps_overlap5)
dim(atlas_aou_overlap_full) # dim = (8705988, 9)
# % matching
8705988/10388380 # 0.8380506
save(atlas_aou_overlap_full, file = paste0(raw_data_path, "atlas/geno/atlas_aou_overlap_full.rda"))
# file for PLINK extraction
atlas_aou_overlap_short = atlas_aou_overlap_full %>% select(rsid)
# output to txt
write.table(atlas_aou_overlap_short, 
            file = paste0(raw_data_path, 'atlas/geno/atlas.aou.overlap.txt'),
            sep = "\t", quote = F, row.names = F, col.names = F)
```

# Part 2. FUMA results cleaning
## 1. Extract all candidate SNPs 
```{r message=FALSE, warning=FALSE}
rm(list = ls())
# lapply(paste('package:', names(sessionInfo()$otherPkgs), sep = ""), 
#        detach, character.only = TRUE, unload = TRUE)
pacman::p_load(tidyverse, GenomicRanges, rtracklayer, bigsnpr, bigreadr, 
               data.table, ieugwasr, MRutils)
raw_data_path = "/Users/Mingzhou/Desktop/Projects/Dementia-prediction/data/"
output_path = "/Users/Mingzhou/Desktop/Projects/Dementia-prediction/output/"
# Source in useful functions
source("/Users/Mingzhou/Desktop/Projects/Dementia-prediction/code/functions.R")
```

### 1) AD Kunkle EUR
```{r}
# Full list of independent significant SNPs + LD SNPs (candidate SNPs, catch as many SNPs as we can from ATLAS)
full_AD_EUR = read.table(file = paste0(raw_data_path, "FUMA_outputs/FUMA_AD_EUR/snps.txt"), 
                         header = T, sep = "\t", fill = T)
lead_AD_EUR = read.table(file = paste0(raw_data_path, "FUMA_outputs/FUMA_AD_EUR/leadsnps.txt"), 
                         header = T, sep = "\t", fill = T)
candi_AD_EUR = full_AD_EUR %>% filter(!is.na(gwasP)) %>% 
  mutate(lead = if_else(rsID %in% lead_AD_EUR$rsID, 1, 0)) %>% 
  mutate(phenotype = "AD_EUR") %>% 
  select(chr, pos, rsID, non_effect_allele, effect_allele, beta, gwasP, IndSigSNP, 
         func, CADD, posMapFilt, eqtlMapFilt, ciMapFilt, lead, phenotype) %>% unique()
dim(candi_AD_EUR) # dim = (3022,15)
```

### 2) AD Kunkle AFR
```{r}
full_AD_AFR = read.table(file = paste0(raw_data_path, "FUMA_outputs/FUMA_AD_AFR/snps.txt"), 
                         header = T, sep = "\t", fill = T)
lead_AD_AFR = read.table(file = paste0(raw_data_path, "FUMA_outputs/FUMA_AD_AFR/leadsnps.txt"), 
                         header = T, sep = "\t", fill = T)
candi_AD_AFR = full_AD_AFR %>% filter(!is.na(gwasP)) %>% 
  mutate(lead = if_else(rsID %in% lead_AD_AFR$rsID, 1, 0)) %>% 
  mutate(phenotype = "AD_AFR") %>% 
  select(chr, pos, rsID, non_effect_allele, effect_allele, beta, gwasP, IndSigSNP, 
         func, CADD, posMapFilt, eqtlMapFilt, ciMapFilt, lead, phenotype) %>% unique()
dim(candi_AD_AFR) # dim = (109,15)
```

### 3) AD Jun trans
```{r}
full_AD_trans = read.table(file = paste0(raw_data_path, "FUMA_outputs/FUMA_AD_trans/snps.txt"), 
                           header = T, sep = "\t", fill = T)
lead_AD_trans = read.table(file = paste0(raw_data_path, "FUMA_outputs/FUMA_AD_trans/leadsnps.txt"), 
                           header = T, sep = "\t", fill = T)
candi_AD_trans = full_AD_trans %>% filter(!is.na(gwasP)) %>% 
  mutate(lead = if_else(rsID %in% lead_AD_trans$rsID, 1, 0)) %>% 
  mutate(phenotype = "AD_trans") %>% 
  select(chr, pos, rsID, non_effect_allele, effect_allele, beta, gwasP, IndSigSNP, 
         func, CADD, posMapFilt, eqtlMapFilt, ciMapFilt, lead, phenotype) %>% unique()
dim(candi_AD_trans) # dim = (1349,15)
```

### 4) PD
```{r}
full_PD = read.table(file = paste0(raw_data_path, "FUMA_outputs/FUMA_PD/snps.txt"), 
                     header = T, sep = "\t", fill = T)
lead_PD = read.table(file = paste0(raw_data_path, "FUMA_outputs/FUMA_PD/leadsnps.txt"), 
                     header = T, sep = "\t", fill = T)
candi_PD = full_PD %>% filter(!is.na(gwasP)) %>% 
  mutate(lead = if_else(rsID %in% lead_PD$rsID, 1, 0)) %>% 
  mutate(phenotype = "PD") %>% 
  select(chr, pos, rsID, non_effect_allele, effect_allele, beta, gwasP, IndSigSNP, 
         func, CADD, posMapFilt, eqtlMapFilt, ciMapFilt, lead, phenotype) %>% unique()
dim(candi_PD) # dim = (7154,15)
```

### 5) PSP
```{r}
full_PSP = read.table(file = paste0(raw_data_path, "FUMA_outputs/FUMA_PSP/snps.txt"), 
                     header = T, sep = "\t", fill = T)
lead_PSP = read.table(file = paste0(raw_data_path, "FUMA_outputs/FUMA_PSP/leadsnps.txt"), 
                     header = T, sep = "\t", fill = T)
candi_PSP = full_PSP %>% filter(!is.na(gwasP)) %>% 
  mutate(lead = if_else(rsID %in% lead_PSP$rsID, 1, 0)) %>% 
  mutate(phenotype = "PSP") %>% 
  select(chr, pos, rsID, non_effect_allele, effect_allele, beta, gwasP, IndSigSNP, 
         func, CADD, posMapFilt, eqtlMapFilt, ciMapFilt, lead, phenotype) %>% unique()
dim(candi_PSP) # dim = (4347,15)
```

### 6) LBD
```{r}
full_LBD = read.table(file = paste0(raw_data_path, "FUMA_outputs/FUMA_LBD/snps.txt"), 
                     header = T, sep = "\t", fill = T)
lead_LBD = read.table(file = paste0(raw_data_path, "FUMA_outputs/FUMA_LBD/leadsnps.txt"), 
                     header = T, sep = "\t", fill = T)
candi_LBD = full_LBD %>% filter(!is.na(gwasP)) %>% 
  mutate(lead = if_else(rsID %in% lead_LBD$rsID, 1, 0)) %>% 
  mutate(phenotype = "LBD") %>% 
  select(chr, pos, rsID, non_effect_allele, effect_allele, beta, gwasP, IndSigSNP, 
         func, CADD, posMapFilt, eqtlMapFilt, ciMapFilt, lead, phenotype) %>% unique()
dim(candi_LBD) # dim = (658,15)
```

### 7) STROKE
```{r}
full_STROKE = read.table(file = paste0(raw_data_path, "FUMA_outputs/FUMA_STROKE/snps.txt"), 
                     header = T, sep = "\t", fill = T)
lead_STROKE = read.table(file = paste0(raw_data_path, "FUMA_outputs/FUMA_STROKE/leadsnps.txt"), 
                     header = T, sep = "\t", fill = T)
candi_STROKE = full_STROKE %>% filter(!is.na(gwasP)) %>% 
  mutate(lead = if_else(rsID %in% lead_STROKE$rsID, 1, 0)) %>% 
  mutate(phenotype = "STROKE") %>% 
  select(chr, pos, rsID, non_effect_allele, effect_allele, beta, gwasP, IndSigSNP, 
         func, CADD, posMapFilt, eqtlMapFilt, ciMapFilt, lead, phenotype) %>% unique()
dim(candi_STROKE) # dim = (1589,15)
```

## 2. Cleaning (liftover + annotation)
```{r}
candi_full = rbind(candi_AD_EUR, candi_AD_AFR, candi_AD_trans, candi_PD, 
                   candi_PSP, candi_LBD, candi_STROKE) %>% as.data.frame() %>% 
  mutate(index = seq(1, nrow(.)))
# liftover
chain_file = import.chain("hg19ToHg38.over.chain")
# Prepare Genomic Ranges
hg19.gr = GRanges(
  seqname = paste("chr", candi_full$chr, sep = ""),
  ranges = IRanges(start = as.numeric(candi_full$pos), 
                   end = as.numeric(candi_full$pos)),
  snp.name = candi_full$rsID)
candi_hg38_liftover = as.data.frame(liftOver(hg19.gr, chain_file)) %>% 
  mutate(chr = as.integer(substring(seqnames, 4, last = 1000000L))) %>% 
  select(group, chr, start) %>% unique() %>% drop_na()
liftover_full = candi_full %>% 
  left_join(candi_hg38_liftover, by = c("index" = "group", "chr" = "chr")) %>% 
  select(chr, start, rsID, non_effect_allele, effect_allele, beta, gwasP, IndSigSNP, 
         func, CADD, posMapFilt, eqtlMapFilt, ciMapFilt, lead, phenotype) %>% 
  dplyr::rename(pos = start, a0 = non_effect_allele, a1 = effect_allele) %>% 
  filter(!is.na(chr) & !is.na(pos)) %>% unique()
dim(liftover_full) # dim = 18199,15
```

## 3. Extract and clean ATLAS data -- Server step
```{r}
# Load ATLAS AOU overlaps
load(file = paste0(raw_data_path, "atlas/geno/atlas_aou_overlap_full.rda"))
atlas_aou_overlap_short = atlas_aou_overlap_full %>% 
  select(rsid, chr, start, ref, alt)
dim(atlas_aou_overlap_short) # dim = 8705988,5
# Check overlap with FUMA
liftover_1 = liftover_full %>% dplyr::rename("chr_num" = "chr") %>% 
  mutate(chr_pos_name1 = paste0("chr", chr_num, ":", pos, ":", a0, ":", a1)) %>% 
  inner_join(atlas_aou_overlap_short, by = c("chr_pos_name1" = "rsid")) %>% 
  dplyr::rename("chr_pos_name" = "chr_pos_name1")
liftover_2 = liftover_full %>% dplyr::rename("chr_num" = "chr") %>% 
  mutate(chr_pos_name2 = paste0("chr", chr_num, ":", pos, ":", a1, ":", a0)) %>% 
  inner_join(atlas_aou_overlap_short, by = c("chr_pos_name2" = "rsid")) %>% 
  dplyr::rename("chr_pos_name" = "chr_pos_name2")
overlap_fuma = rbind(liftover_1, liftover_2) %>% unique()
dim(overlap_fuma) # dim = 8727,20
# Save extract_fuma
extract_fuma = overlap_fuma %>% pull(chr_pos_name) %>% unique()
length(extract_fuma) # length = 7393
save(overlap_fuma, file = paste0(raw_data_path, "FUMA_outputs/overlap_fuma.rda"))
# Write output for snp extraction on server
write.table(extract_fuma, 
            file = paste0(raw_data_path, "FUMA_outputs/extract_fuma.txt"),
            sep = "\t", quote = F, row.names = F, col.names = F)
```

```{r}
# Read in the QCed genotype and sumstats files + preprocess the bed file (only need to do once for each data set)
geno_file = paste0(raw_data_path, 'atlas/geno/fuma.raw')
snp_readBed(paste0(geno_file, '.bed'))
obj.bigSNP = snp_attach(paste0(geno_file, '.rds'))
# extract the SNP information from the genotype
map = obj.bigSNP$map[-3]
names(map) = c("chr", "rsid", "pos", "a1", "a0")
# Rename rsid to keep consistent with bim file
map = map %>% mutate(RSID = paste0("chr", chr, ":", pos, ":", a0, ":", a1)) %>% 
  select(chr, RSID, pos, a1, a0) %>% dplyr::rename("rsid" = "RSID")
# Get fam order
fam.order = as.data.table(obj.bigSNP$fam)
# get genotype info
genotype = obj.bigSNP$genotypes
genotype_df = as.matrix(genotype[])[,] %>% as.data.frame()
colnames(genotype_df) = map$rsid
# check missing patterns
# gg_miss_upset(genotype_df)
```

```{r}
# The above SNPs have a missing >10000, we decide to exclude those SNPs (N = 343)
missing_info = genotype_df %>% 
  summarise_all(function(x) sum(is.na(x))) %>%  
  gather(variable, missing_count) 
columns_with_missing = missing_info %>% 
  filter(missing_count > 10000) %>% pull(variable)
genotype_df_noNA = genotype_df %>% select(-all_of(columns_with_missing)) 
dim(genotype_df_noNA) # 57483,7050
genotype_df_fam = cbind(fam.order[,1:2], genotype_df_noNA) %>% 
  as.data.frame() %>% drop_na()
colnames(genotype_df_fam)[1:2] = c("FID", "IID")
dim(genotype_df_fam) # 57483,7052
# also need to update map file
map_noNA = map %>% filter(rsid %!in% columns_with_missing)
dim(map_noNA) # 7050,5
save(map_noNA, file = paste0(raw_data_path, "modeling/map_noNA_full_SNP.rda"))
save(genotype_df_fam, file = paste0(raw_data_path, "modeling/freeze60k_geno_freq_full_SNP.rda"))
```


# Part 3. Finalize datasets
## 1. AD EUR
```{r}
snp_AD_EUR = overlap_fuma %>% filter(phenotype == "AD_EUR") %>% 
  mutate(ind_sig = if_else(rsID == IndSigSNP, 1, 0)) %>% filter(!is.na(beta)) %>% 
  mutate(map_yes = case_when(
    posMapFilt == 1 | eqtlMapFilt == 1 | ciMapFilt == 1 ~ 1,
    TRUE ~ 0
  )) %>% select(chr_num, pos, a1, a0, beta, gwasP, rsID, IndSigSNP, 
                ind_sig, lead, map_yes, CADD) %>% unique() %>% 
  dplyr::rename(chr = chr_num)
# Reverse beta if necessary by performing SNP matching
info_snp_AD_EUR = snp_match(snp_AD_EUR, map_noNA) # 1744 variants have been matched
```

### A. Independent significant SNPs 
```{r}
final_extract1 = info_snp_AD_EUR %>% filter(ind_sig == 1) %>% unique() # N = 45
dim(final_extract1)
final_extract2 = info_snp_AD_EUR %>% filter(ind_sig == 0) %>% 
  filter(IndSigSNP %!in% final_extract1$IndSigSNP) %>% filter(gwasP < 0.05) %>% 
  group_by(IndSigSNP) %>% arrange(gwasP) %>% mutate(order = row_number()) %>% 
  filter(order == 1) %>% select(-order) # N = 31
dim(final_extract2)
indsig_AD_EUR_final = rbind(final_extract1, final_extract2) %>% unique() %>% 
  select(rsid, chr, pos, a1, a0, beta, gwasP, rsID) 
dim(indsig_AD_EUR_final) # dim = (76,8)
# Remap to get final
info_snp = snp_match(indsig_AD_EUR_final, map_noNA) # 76 variants have been matched
indsig_beta_AD_EUR = info_snp$beta
AD_EUR_beta_indsig = info_snp %>% 
  select(rsid, chr, pos, a0, a1, beta, gwasP) %>% 
  dplyr::rename(beta_EUR_indsig = beta)
# Filter genotype info
genotype.targ.indsig = genotype_df_fam %>% dplyr::select(all_of(AD_EUR_beta_indsig$rsid))
genotype.targ.matrix = as.matrix(genotype.targ.indsig[])
# Calculate PRS
pred_prs = genotype.targ.matrix %*% indsig_beta_AD_EUR
pred_prs_final = cbind(pred_prs, genotype_df_fam[, c('FID','IID')])
names(pred_prs_final) = c('AD_PRS_EUR_indsig', 'FID', 'IID')
# Normalization
onekg_EUR = pred_prs_final %>% filter(FID == "EUR")
AD_EUR_1kg_mean = mean(onekg_EUR$AD_PRS_EUR_indsig) # -0.9308247
AD_EUR_1kg_sd = sd(onekg_EUR$AD_PRS_EUR_indsig) # 1.505618
# Apply to ATLAS
AD_EUR_prs_indsig = pred_prs_final %>% 
  filter(FID %!in% c("EUR", "AFR", "AMR", "EAS", "SAS"))
dim(AD_EUR_prs_indsig) # dim = (54935,3)
AD_EUR_prs_indsig$AD_PRS_EUR_indsig_norm = 
  (AD_EUR_prs_indsig$AD_PRS_EUR_indsig - AD_EUR_1kg_mean) / AD_EUR_1kg_sd
AD_EUR_prs_indsig = AD_EUR_prs_indsig %>% 
  select(IID, AD_PRS_EUR_indsig, AD_PRS_EUR_indsig_norm) %>% 
  dplyr::rename(UniqueSampleId = IID)
```

### B. Lead SNPs
```{r}
final_extract = info_snp_AD_EUR %>% filter(lead == 1) %>% unique() # N = 27
dim(final_extract)
lead_AD_EUR_final = final_extract %>% 
  select(rsid, chr, pos, a1, a0, beta, gwasP, rsID)
dim(lead_AD_EUR_final) # dim = (27,8)
# Remap to get final
info_snp = snp_match(lead_AD_EUR_final, map_noNA) # 27 variants have been matched
lead_beta_AD_EUR = info_snp$beta
AD_EUR_beta_lead = info_snp %>% 
  select(rsid, chr, pos, a0, a1, beta, gwasP) %>% 
  dplyr::rename(beta_EUR_lead = beta)
# Filter genotype info
genotype.targ.lead = genotype_df_fam %>% select(all_of(AD_EUR_beta_lead$rsid))
genotype.targ.matrix = as.matrix(genotype.targ.lead[])
# Calculate PRS
pred_prs = genotype.targ.matrix %*% lead_beta_AD_EUR
pred_prs_final = cbind(pred_prs, genotype_df_fam[, c('FID','IID')])
names(pred_prs_final) = c('AD_PRS_EUR_lead', 'FID', 'IID')
# Normalization
onekg_EUR = pred_prs_final %>% filter(FID == "EUR")
AD_EUR_1kg_mean = mean(onekg_EUR$AD_PRS_EUR_lead) # 0.2792383
AD_EUR_1kg_sd = sd(onekg_EUR$AD_PRS_EUR_lead) # 0.8502298
# Apply to ATLAS
AD_EUR_prs_lead = pred_prs_final %>% 
  filter(FID %!in% c("EUR", "AFR", "AMR", "EAS", "SAS"))
dim(AD_EUR_prs_lead) # dim = (54935,3)
AD_EUR_prs_lead$AD_PRS_EUR_lead_norm = 
  (AD_EUR_prs_lead$AD_PRS_EUR_lead - AD_EUR_1kg_mean) / AD_EUR_1kg_sd
AD_EUR_prs_lead = AD_EUR_prs_lead %>% 
  select(IID, AD_PRS_EUR_lead, AD_PRS_EUR_lead_norm) %>% 
  dplyr::rename(UniqueSampleId = IID)
```

### C. Mapped SNPs (independent of LD)
```{r}
final_extract = info_snp_AD_EUR %>% filter(map_yes == 1) %>%
  group_by(IndSigSNP) %>% arrange(desc(CADD)) %>%
  mutate(order = row_number()) %>% filter(order == 1) %>%
  select(-order) %>% ungroup()
map_AD_EUR_final = final_extract %>% 
  select(rsid, chr, pos, a1, a0, beta, gwasP, rsID)
dim(map_AD_EUR_final) # dim = (75,8)
# Remap to get final
info_snp = snp_match(map_AD_EUR_final, map_noNA) # 75 variants have been matched
map_beta_AD_EUR = info_snp$beta
AD_EUR_beta_map = info_snp %>% 
  select(rsid, chr, pos, a0, a1, beta, gwasP) %>% 
  dplyr::rename(beta_EUR_map = beta)
# Filter genotype info
genotype.targ.map = genotype_df_fam %>% select(all_of(AD_EUR_beta_map$rsid))
genotype.targ.matrix = as.matrix(genotype.targ.map[])
# Calculate PRS
pred_prs = genotype.targ.matrix %*% map_beta_AD_EUR
pred_prs_final = cbind(pred_prs, genotype_df_fam[, c('FID','IID')])
names(pred_prs_final) = c('AD_PRS_EUR_map', 'FID', 'IID')
# Normalization
onekg_EUR = pred_prs_final %>% filter(FID == "EUR")
AD_EUR_1kg_mean = mean(onekg_EUR$AD_PRS_EUR_map) # -0.0667567
AD_EUR_1kg_sd = sd(onekg_EUR$AD_PRS_EUR_map) # 1.134523
# Apply to ATLAS
AD_EUR_prs_map = pred_prs_final %>% 
  filter(FID %!in% c("EUR", "AFR", "AMR", "EAS", "SAS"))
dim(AD_EUR_prs_map) # dim = (54935,3)
AD_EUR_prs_map$AD_PRS_EUR_map_norm = 
  (AD_EUR_prs_map$AD_PRS_EUR_map - AD_EUR_1kg_mean) / AD_EUR_1kg_sd
AD_EUR_prs_map = AD_EUR_prs_map %>% 
  select(IID, AD_PRS_EUR_map, AD_PRS_EUR_map_norm) %>% 
  dplyr::rename(UniqueSampleId = IID)
```

### D. Mapped SNPs (independent of LD) with CADD >= 12.37
```{r}
final_extract = info_snp_AD_EUR %>% filter(map_yes == 1) %>%
  filter(CADD >= 12.37) %>% 
  group_by(IndSigSNP) %>% arrange(desc(CADD)) %>%
  mutate(order = row_number()) %>% filter(order == 1) %>%
  select(-order) %>% ungroup()
map_CADD_AD_EUR_final = final_extract %>% 
  select(rsid, chr, pos, a1, a0, beta, gwasP, rsID)
dim(map_CADD_AD_EUR_final) # dim = (54,8)
# Remap to get final
info_snp = snp_match(map_CADD_AD_EUR_final, map_noNA) # 45 variants have been matched
map_CADD_beta_AD_EUR = info_snp$beta
AD_EUR_beta_map_CADD = info_snp %>% 
  select(rsid, chr, pos, a0, a1, beta, gwasP) %>% 
  dplyr::rename(beta_EUR_map_CADD = beta)
# Filter genotype info
genotype.targ.map_CADD = genotype_df_fam %>% select(all_of(AD_EUR_beta_map_CADD$rsid))
genotype.targ.matrix = as.matrix(genotype.targ.map_CADD[])
# Calculate PRS
pred_prs = genotype.targ.matrix %*% map_CADD_beta_AD_EUR
pred_prs_final = cbind(pred_prs, genotype_df_fam[, c('FID','IID')])
names(pred_prs_final) = c('AD_PRS_EUR_map_CADD', 'FID', 'IID')
# Normalization
onekg_EUR = pred_prs_final %>% filter(FID == "EUR")
AD_EUR_1kg_mean = mean(onekg_EUR$AD_PRS_EUR_map_CADD)
AD_EUR_1kg_sd = sd(onekg_EUR$AD_PRS_EUR_map_CADD)
# Apply to ATLAS
AD_EUR_prs_map_CADD = pred_prs_final %>% 
  filter(FID %!in% c("EUR", "AFR", "AMR", "EAS", "SAS"))
dim(AD_EUR_prs_map_CADD) # dim = (54935,3)
AD_EUR_prs_map_CADD$AD_PRS_EUR_map_CADD_norm = 
  (AD_EUR_prs_map_CADD$AD_PRS_EUR_map_CADD - AD_EUR_1kg_mean) / AD_EUR_1kg_sd
AD_EUR_prs_map_CADD = AD_EUR_prs_map_CADD %>% 
  select(IID, AD_PRS_EUR_map_CADD, AD_PRS_EUR_map_CADD_norm) %>% 
  dplyr::rename(UniqueSampleId = IID)
```

### Combine all AD EUR PRSs
```{r}
AD_Kunkle_EUR_prs = AD_EUR_prs_indsig %>% inner_join(AD_EUR_prs_lead) %>% 
  inner_join(AD_EUR_prs_map) %>% inner_join(AD_EUR_prs_map_CADD) 
AD_Kunkle_EUR_beta = AD_EUR_beta_indsig %>% full_join(AD_EUR_beta_lead) %>% 
  full_join(AD_EUR_beta_map) %>% full_join(AD_EUR_beta_map_CADD) %>% select(-gwasP)
save(AD_Kunkle_EUR_prs, file = paste0(raw_data_path, "prs/raw/AD_Kunkle_EUR_prs.rda"))
save(AD_Kunkle_EUR_beta, file = paste0(raw_data_path, "prs/raw/AD_Kunkle_EUR_beta.rda"))
```

## 2. AD AFR
```{r}
snp_AD_AFR = overlap_fuma %>% filter(phenotype == "AD_AFR") %>% 
  mutate(ind_sig = if_else(rsID == IndSigSNP, 1, 0)) %>% filter(!is.na(beta)) %>% 
  mutate(map_yes = case_when(
    posMapFilt == 1 | eqtlMapFilt == 1 | ciMapFilt == 1 ~ 1,
    TRUE ~ 0
  )) %>% select(chr_num, pos, a1, a0, beta, gwasP, rsID, IndSigSNP, 
                ind_sig, lead, map_yes, CADD) %>% unique() %>% 
  dplyr::rename(chr = chr_num)
# Reverse beta if necessary by performing SNP matching
info_snp_AD_AFR = snp_match(snp_AD_AFR, map_noNA) # 80 variants have been matched
```

### A. Independent significant SNPs 
```{r}
final_extract1 = info_snp_AD_AFR %>% filter(ind_sig == 1) %>% unique() # N = 9
dim(final_extract1)
final_extract2 = info_snp_AD_AFR %>% filter(ind_sig == 0) %>% 
  filter(IndSigSNP %!in% final_extract1$IndSigSNP) %>% filter(gwasP < 0.05) %>% 
  group_by(IndSigSNP) %>% arrange(gwasP) %>% mutate(order = row_number()) %>% 
  filter(order == 1) %>% select(-order) 
dim(final_extract2) # N = 2
indsig_AD_AFR_final = rbind(final_extract1, final_extract2) %>% unique() %>% 
  select(rsid, chr, pos, a1, a0, beta, gwasP, rsID)
dim(indsig_AD_AFR_final) # dim = (11,8)
# Remap to get final
info_snp = snp_match(indsig_AD_AFR_final, map_noNA) # 11 variants have been matched
indsig_beta_AD_AFR = info_snp$beta
AD_AFR_beta_indsig = info_snp %>% 
  select(rsid, chr, pos, a0, a1, beta, gwasP) %>% 
  dplyr::rename(beta_AFR_indsig = beta)
# Save this for future use
save(AD_AFR_beta_indsig, 
     file = paste0(raw_data_path, "prs/mod/AD_AFR_beta_indsig.rda"))
# Filter genotype info
genotype.targ.indsig = genotype_df_fam %>% select(all_of(AD_AFR_beta_indsig$rsid))
genotype.targ.matrix = as.matrix(genotype.targ.indsig[])
# Calculate PRS
pred_prs = genotype.targ.matrix %*% indsig_beta_AD_AFR
pred_prs_final = cbind(pred_prs, genotype_df_fam[, c('FID','IID')])
names(pred_prs_final) = c('AD_PRS_AFR_indsig', 'FID', 'IID')
# Normalization
onekg_EUR = pred_prs_final %>% filter(FID == "EUR")
AD_EUR_1kg_mean = mean(onekg_EUR$AD_PRS_AFR_indsig) # 0.3909557
AD_EUR_1kg_sd = sd(onekg_EUR$AD_PRS_AFR_indsig) # 1.171444
# Apply to ATLAS
AD_AFR_prs_indsig = pred_prs_final %>% 
  filter(FID %!in% c("EUR", "AFR", "AMR", "EAS", "SAS"))
dim(AD_AFR_prs_indsig) # dim = (54935,3)
AD_AFR_prs_indsig$AD_PRS_AFR_indsig_norm = 
  (AD_AFR_prs_indsig$AD_PRS_AFR_indsig - AD_EUR_1kg_mean) / AD_EUR_1kg_sd
AD_AFR_prs_indsig = AD_AFR_prs_indsig %>% 
  select(IID, AD_PRS_AFR_indsig, AD_PRS_AFR_indsig_norm) %>% 
  dplyr::rename(UniqueSampleId = IID)
```

### B. Lead SNPs
```{r}
final_extract = info_snp_AD_AFR %>% filter(lead == 1) %>% unique() # N = 4
dim(final_extract)
lead_AD_AFR_final = final_extract %>% 
  select(rsid, chr, pos, a1, a0, beta, gwasP, rsID)
dim(lead_AD_AFR_final) # dim = (4,8)
# Remap to get final
info_snp = snp_match(lead_AD_AFR_final, map_noNA) # 4 variants have been matched
lead_beta_AD_AFR = info_snp$beta
AD_AFR_beta_lead = info_snp %>% 
  select(rsid, chr, pos, a0, a1, beta, gwasP) %>% 
  dplyr::rename(beta_AFR_lead = beta)
# Filter genotype info
genotype.targ.lead = genotype_df_fam %>% select(all_of(AD_AFR_beta_lead$rsid))
genotype.targ.matrix = as.matrix(genotype.targ.lead[])
# Calculate PRS
pred_prs = genotype.targ.matrix %*% lead_beta_AD_AFR
pred_prs_final = cbind(pred_prs, genotype_df_fam[, c('FID','IID')])
names(pred_prs_final) = c('AD_PRS_AFR_lead', 'FID', 'IID')
# Normalization
onekg_EUR = pred_prs_final %>% filter(FID == "EUR")
AD_EUR_1kg_mean = mean(onekg_EUR$AD_PRS_AFR_lead) # 0.1599015
AD_EUR_1kg_sd = sd(onekg_EUR$AD_PRS_AFR_lead) # 0.4512958
# Apply to ATLAS
AD_AFR_prs_lead = pred_prs_final %>% 
  filter(FID %!in% c("EUR", "AFR", "AMR", "EAS", "SAS"))
dim(AD_AFR_prs_lead) # dim = (54935,3)
AD_AFR_prs_lead$AD_PRS_AFR_lead_norm = 
  (AD_AFR_prs_lead$AD_PRS_AFR_lead - AD_EUR_1kg_mean) / AD_EUR_1kg_sd
AD_AFR_prs_lead = AD_AFR_prs_lead %>% 
  select(IID, AD_PRS_AFR_lead, AD_PRS_AFR_lead_norm) %>% 
  dplyr::rename(UniqueSampleId = IID)
```

### C. Mapped SNPs
```{r}
final_extract = info_snp_AD_AFR %>% filter(map_yes == 1) %>% 
  group_by(IndSigSNP) %>% arrange(desc(CADD)) %>% 
  mutate(order = row_number()) %>% filter(order == 1) %>% 
  select(-order) %>% ungroup()
dim(final_extract)
map_AD_AFR_final = final_extract %>% 
  select(rsid, chr, pos, a1, a0, beta, gwasP, rsID)
dim(map_AD_AFR_final) # dim = (11,8)
# Remap to get final
info_snp = snp_match(map_AD_AFR_final, map_noNA) # 11 variants have been matched
map_beta_AD_AFR = info_snp$beta
AD_AFR_beta_map = info_snp %>% 
  select(rsid, chr, pos, a0, a1, beta, gwasP) %>% 
  dplyr::rename(beta_AFR_map = beta)
# Filter genotype info
genotype.targ.map = genotype_df_fam %>% select(all_of(AD_AFR_beta_map$rsid))
genotype.targ.matrix = as.matrix(genotype.targ.map[])
# Calculate PRS
pred_prs = genotype.targ.matrix %*% map_beta_AD_AFR
pred_prs_final = cbind(pred_prs, genotype_df_fam[, c('FID','IID')])
names(pred_prs_final) = c('AD_PRS_AFR_map', 'FID', 'IID')
# Normalization
onekg_EUR = pred_prs_final %>% filter(FID == "EUR")
AD_EUR_1kg_mean = mean(onekg_EUR$AD_PRS_AFR_map) # 0.1592197
AD_EUR_1kg_sd = sd(onekg_EUR$AD_PRS_AFR_map) # 0.752526
# Apply to ATLAS
AD_AFR_prs_map = pred_prs_final %>% 
  filter(FID %!in% c("EUR", "AFR", "AMR", "EAS", "SAS"))
dim(AD_AFR_prs_map) # dim = (54935,3)
AD_AFR_prs_map$AD_PRS_AFR_map_norm = 
  (AD_AFR_prs_map$AD_PRS_AFR_map - AD_EUR_1kg_mean) / AD_EUR_1kg_sd
AD_AFR_prs_map = AD_AFR_prs_map %>% 
  select(IID, AD_PRS_AFR_map, AD_PRS_AFR_map_norm) %>% 
  dplyr::rename(UniqueSampleId = IID)
```

### D. Mapped SNPs (independent of LD) with CADD >= 12.37
```{r}
final_extract = info_snp_AD_AFR %>% filter(map_yes == 1) %>%
  filter(CADD >= 12.37) %>% 
  group_by(IndSigSNP) %>% arrange(desc(CADD)) %>%
  mutate(order = row_number()) %>% filter(order == 1) %>%
  select(-order) %>% ungroup()
map_CADD_AD_AFR_final = final_extract %>% 
  select(rsid, chr, pos, a1, a0, beta, gwasP, rsID)
dim(map_CADD_AD_AFR_final) # dim = (5,8)
# Remap to get final
info_snp = snp_match(map_CADD_AD_AFR_final, map_noNA) # 5 variants have been matched
map_CADD_beta_AD_AFR = info_snp$beta
AD_AFR_beta_map_CADD = info_snp %>% 
  select(rsid, chr, pos, a0, a1, beta, gwasP) %>% 
  dplyr::rename(beta_AFR_map_CADD = beta)
# Filter genotype info
genotype.targ.map_CADD = genotype_df_fam %>% select(all_of(AD_AFR_beta_map_CADD$rsid))
genotype.targ.matrix = as.matrix(genotype.targ.map_CADD[])
# Calculate PRS
pred_prs = genotype.targ.matrix %*% map_CADD_beta_AD_AFR
pred_prs_final = cbind(pred_prs, genotype_df_fam[, c('FID','IID')])
names(pred_prs_final) = c('AD_PRS_AFR_map_CADD', 'FID', 'IID')
# Normalization
onekg_EUR = pred_prs_final %>% filter(FID == "EUR")
AD_EUR_1kg_mean = mean(onekg_EUR$AD_PRS_AFR_map_CADD)
AD_EUR_1kg_sd = sd(onekg_EUR$AD_PRS_AFR_map_CADD)
# Apply to ATLAS
AD_AFR_prs_map_CADD = pred_prs_final %>% 
  filter(FID %!in% c("EUR", "AFR", "AMR", "EAS", "SAS"))
dim(AD_AFR_prs_map_CADD) # dim = (54935,3)
AD_AFR_prs_map_CADD$AD_PRS_AFR_map_CADD_norm = 
  (AD_AFR_prs_map_CADD$AD_PRS_AFR_map_CADD - AD_EUR_1kg_mean) / AD_EUR_1kg_sd
AD_AFR_prs_map_CADD = AD_AFR_prs_map_CADD %>% 
  select(IID, AD_PRS_AFR_map_CADD, AD_PRS_AFR_map_CADD_norm) %>% 
  dplyr::rename(UniqueSampleId = IID)
```

### Combine all AD AFR PRSs
```{r}
AD_Kunkle_AFR_prs = AD_AFR_prs_indsig %>% inner_join(AD_AFR_prs_lead) %>% 
  inner_join(AD_AFR_prs_map) %>% inner_join(AD_AFR_prs_map_CADD)
AD_Kunkle_AFR_beta = AD_AFR_beta_indsig %>% full_join(AD_AFR_beta_lead) %>% 
  full_join(AD_AFR_beta_map) %>% full_join(AD_AFR_beta_map_CADD) %>% select(-gwasP)
save(AD_Kunkle_AFR_prs, file = paste0(raw_data_path, "prs/raw/AD_Kunkle_AFR_prs.rda"))
save(AD_Kunkle_AFR_beta, file = paste0(raw_data_path, "prs/raw/AD_Kunkle_AFR_beta.rda"))
```

## 3. AD trans
```{r}
snp_AD_trans = overlap_fuma %>% filter(phenotype == "AD_trans") %>% 
  mutate(ind_sig = if_else(rsID == IndSigSNP, 1, 0)) %>% filter(!is.na(beta)) %>% 
  mutate(map_yes = case_when(
    posMapFilt == 1 | eqtlMapFilt == 1 | ciMapFilt == 1 ~ 1,
    TRUE ~ 0
  )) %>% select(chr_num, pos, a1, a0, beta, gwasP, rsID, IndSigSNP, 
                ind_sig, lead, map_yes, CADD) %>% unique() %>% 
  dplyr::rename(chr = chr_num)
# Reverse beta if necessary by performing SNP matching
info_snp_AD_trans = snp_match(snp_AD_trans, map_noNA) # 760 variants have been matched
```

### A. Independent significant SNPs 
```{r}
final_extract1 = info_snp_AD_trans %>% filter(ind_sig == 1) %>% unique() # N = 38
dim(final_extract1)
final_extract2 = info_snp_AD_trans %>% filter(ind_sig == 0) %>% 
  filter(IndSigSNP %!in% final_extract1$IndSigSNP) %>% filter(gwasP < 0.05) %>% 
  group_by(IndSigSNP) %>% arrange(gwasP) %>% mutate(order = row_number()) %>% 
  filter(order == 1) %>% select(-order) 
dim(final_extract2)
indsig_AD_trans_final = rbind(final_extract1, final_extract2) %>% unique() %>% 
  select(rsid, chr, pos, a1, a0, beta, gwasP, rsID)
dim(indsig_AD_trans_final) # dim = (54,8)
# Remap to get final
info_snp = snp_match(indsig_AD_trans_final, map_noNA) # 54 variants have been matched
indsig_beta_AD_trans = info_snp$beta
AD_trans_beta_indsig = info_snp %>% 
  select(rsid, chr, pos, a0, a1, beta, gwasP) %>% 
  dplyr::rename(beta_trans_indsig = beta)
# Filter genotype info
genotype.targ.indsig = genotype_df_fam %>% select(all_of(AD_trans_beta_indsig$rsid))
genotype.targ.matrix = as.matrix(genotype.targ.indsig[])
# Calculate PRS
pred_prs = genotype.targ.matrix %*% indsig_beta_AD_trans
pred_prs_final = cbind(pred_prs, genotype_df_fam[, c('FID','IID')])
names(pred_prs_final) = c('AD_PRS_trans_indsig', 'FID', 'IID')
# Normalization
onekg_EUR = pred_prs_final %>% filter(FID == "EUR")
AD_EUR_1kg_mean = mean(onekg_EUR$AD_PRS_trans_indsig) # -0.04261034
AD_EUR_1kg_sd = sd(onekg_EUR$AD_PRS_trans_indsig) # 2.248376
# Apply to ATLAS
AD_trans_prs_indsig = pred_prs_final %>% 
  filter(FID %!in% c("EUR", "AFR", "AMR", "EAS", "SAS"))
dim(AD_trans_prs_indsig) # dim = (54935,3)
AD_trans_prs_indsig$AD_PRS_trans_indsig_norm = 
  (AD_trans_prs_indsig$AD_PRS_trans_indsig - AD_EUR_1kg_mean) / AD_EUR_1kg_sd
AD_trans_prs_indsig = AD_trans_prs_indsig %>% 
  select(IID, AD_PRS_trans_indsig, AD_PRS_trans_indsig_norm) %>% 
  dplyr::rename(UniqueSampleId = IID)
```

### B. Lead SNPs
```{r}
final_extract = info_snp_AD_trans %>% filter(lead == 1) %>% unique() # N = 25
dim(final_extract)
lead_AD_trans_final = final_extract %>% 
  select(rsid, chr, pos, a1, a0, beta, gwasP, rsID)
dim(lead_AD_trans_final) # dim = (25,8)
# Remap to get final
info_snp = snp_match(lead_AD_trans_final, map_noNA) # 25 variants have been matched
lead_beta_AD_trans = info_snp$beta
AD_trans_beta_lead = info_snp %>% 
  select(rsid, chr, pos, a0, a1, beta, gwasP) %>% 
  dplyr::rename(beta_trans_lead = beta)
# Filter genotype info
genotype.targ.lead = genotype_df_fam %>% select(all_of(AD_trans_beta_lead$rsid))
genotype.targ.matrix = as.matrix(genotype.targ.lead[])
# Calculate PRS
pred_prs = genotype.targ.matrix %*% lead_beta_AD_trans
pred_prs_final = cbind(pred_prs, genotype_df_fam[, c('FID','IID')])
names(pred_prs_final) = c('AD_PRS_trans_lead', 'FID', 'IID')
# Normalization
onekg_EUR = pred_prs_final %>% filter(FID == "EUR")
AD_EUR_1kg_mean = mean(onekg_EUR$AD_PRS_trans_lead) # 0.3146927
AD_EUR_1kg_sd = sd(onekg_EUR$AD_PRS_trans_lead) # 1.382659
# Apply to ATLAS
AD_trans_prs_lead = pred_prs_final %>% 
  filter(FID %!in% c("EUR", "AFR", "AMR", "EAS", "SAS"))
dim(AD_trans_prs_lead) # dim = (54935,3)
AD_trans_prs_lead$AD_PRS_trans_lead_norm = 
  (AD_trans_prs_lead$AD_PRS_trans_lead - AD_EUR_1kg_mean) / AD_EUR_1kg_sd
AD_trans_prs_lead = AD_trans_prs_lead %>% 
  select(IID, AD_PRS_trans_lead, AD_PRS_trans_lead_norm) %>% 
  dplyr::rename(UniqueSampleId = IID)
```

### C. Mapped SNPs
```{r}
final_extract = info_snp_AD_trans %>% filter(map_yes == 1) %>% 
  group_by(IndSigSNP) %>% arrange(desc(CADD)) %>% 
  mutate(order = row_number()) %>% filter(order == 1) %>% 
  select(-order) %>% ungroup()
map_AD_trans_final = final_extract %>% 
  select(rsid, chr, pos, a1, a0, beta, gwasP, rsID)
dim(map_AD_trans_final) # dim = (54,8)
# Remap to get final
info_snp = snp_match(map_AD_trans_final, map_noNA) # 54 variants have been matched
map_beta_AD_trans = info_snp$beta
AD_trans_beta_map = info_snp %>% 
  select(rsid, chr, pos, a0, a1, beta, gwasP) %>% 
  dplyr::rename(beta_trans_map = beta)
# Filter genotype info
genotype.targ.map = genotype_df_fam %>% select(all_of(AD_trans_beta_map$rsid))
genotype.targ.matrix = as.matrix(genotype.targ.map[])
# Calculate PRS
pred_prs = genotype.targ.matrix %*% map_beta_AD_trans
pred_prs_final = cbind(pred_prs, genotype_df_fam[, c('FID','IID')])
names(pred_prs_final) = c('AD_PRS_trans_map', 'FID', 'IID')
# Normalization
onekg_EUR = pred_prs_final %>% filter(FID == "EUR")
AD_EUR_1kg_mean = mean(onekg_EUR$AD_PRS_trans_map) # 0.2465559
AD_EUR_1kg_sd = sd(onekg_EUR$AD_PRS_trans_map) # 1.968068
# Apply to ATLAS
AD_trans_prs_map = pred_prs_final %>% 
  filter(FID %!in% c("EUR", "AFR", "AMR", "EAS", "SAS"))
dim(AD_trans_prs_map) # dim = (54935,3)
AD_trans_prs_map$AD_PRS_trans_map_norm = 
  (AD_trans_prs_map$AD_PRS_trans_map - AD_EUR_1kg_mean) / AD_EUR_1kg_sd
AD_trans_prs_map = AD_trans_prs_map %>% 
  select(IID, AD_PRS_trans_map, AD_PRS_trans_map_norm) %>% 
  dplyr::rename(UniqueSampleId = IID)
```

### D. Mapped SNPs (independent of LD) with CADD >= 12.37
```{r}
final_extract = info_snp_AD_trans %>% filter(map_yes == 1) %>%
  filter(CADD >= 12.37) %>% 
  group_by(IndSigSNP) %>% arrange(desc(CADD)) %>%
  mutate(order = row_number()) %>% filter(order == 1) %>%
  select(-order) %>% ungroup()
map_CADD_AD_trans_final = final_extract %>% 
  select(rsid, chr, pos, a1, a0, beta, gwasP, rsID)
dim(map_CADD_AD_trans_final) # dim = (17,8)
# Remap to get final
info_snp = snp_match(map_CADD_AD_trans_final, map_noNA) # 17 variants have been matched
map_CADD_beta_AD_trans = info_snp$beta
AD_trans_beta_map_CADD = info_snp %>% 
  select(rsid, chr, pos, a0, a1, beta, gwasP) %>% 
  dplyr::rename(beta_trans_map_CADD = beta)
# Filter genotype info
genotype.targ.map_CADD = genotype_df_fam %>% select(all_of(AD_trans_beta_map_CADD$rsid))
genotype.targ.matrix = as.matrix(genotype.targ.map_CADD[])
# Calculate PRS
pred_prs = genotype.targ.matrix %*% map_CADD_beta_AD_trans
pred_prs_final = cbind(pred_prs, genotype_df_fam[, c('FID','IID')])
names(pred_prs_final) = c('AD_PRS_trans_map_CADD', 'FID', 'IID')
# Normalization
onekg_EUR = pred_prs_final %>% filter(FID == "EUR")
AD_EUR_1kg_mean = mean(onekg_EUR$AD_PRS_trans_map_CADD)
AD_EUR_1kg_sd = sd(onekg_EUR$AD_PRS_trans_map_CADD)
# Apply to ATLAS
AD_trans_prs_map_CADD = pred_prs_final %>% 
  filter(FID %!in% c("EUR", "AFR", "AMR", "EAS", "SAS"))
dim(AD_trans_prs_map_CADD) # dim = (54935,3)
AD_trans_prs_map_CADD$AD_PRS_trans_map_CADD_norm = 
  (AD_trans_prs_map_CADD$AD_PRS_trans_map_CADD - AD_EUR_1kg_mean) / AD_EUR_1kg_sd
AD_trans_prs_map_CADD = AD_trans_prs_map_CADD %>% 
  select(IID, AD_PRS_trans_map_CADD, AD_PRS_trans_map_CADD_norm) %>% 
  dplyr::rename(UniqueSampleId = IID)
```

### Combine all AD trans PRSs
```{r}
AD_Jun_trans_prs = AD_trans_prs_indsig %>% inner_join(AD_trans_prs_lead) %>% 
  inner_join(AD_trans_prs_map) %>% inner_join(AD_trans_prs_map_CADD)
AD_Jun_trans_beta = AD_trans_beta_indsig %>% full_join(AD_trans_beta_lead) %>% 
  full_join(AD_trans_beta_map) %>% full_join(AD_trans_beta_map_CADD) %>% select(-gwasP)
save(AD_Jun_trans_prs, file = paste0(raw_data_path, "prs/raw/AD_Jun_trans_prs.rda"))
save(AD_Jun_trans_beta, file = paste0(raw_data_path, "prs/raw/AD_Jun_trans_beta.rda"))
```

## 4. LBD
```{r}
snp_LBD = overlap_fuma %>% filter(phenotype == "LBD") %>% 
  mutate(ind_sig = if_else(rsID == IndSigSNP, 1, 0)) %>% filter(!is.na(beta)) %>% 
  mutate(map_yes = case_when(
    posMapFilt == 1 | eqtlMapFilt == 1 | ciMapFilt == 1 ~ 1,
    TRUE ~ 0
  )) %>% select(chr_num, pos, a1, a0, beta, gwasP, rsID, IndSigSNP, 
                ind_sig, lead, map_yes, CADD) %>% unique() %>% 
  dplyr::rename(chr = chr_num)
# Reverse beta if necessary by performing SNP matching
info_snp_LBD = snp_match(snp_LBD, map_noNA) # 356 variants have been matched
```

### A. Independent significant SNPs 
```{r}
final_extract1 = info_snp_LBD %>% filter(ind_sig == 1) %>% unique() # N = 8
dim(final_extract1)
final_extract2 = info_snp_LBD %>% filter(ind_sig == 0) %>% 
  filter(IndSigSNP %!in% final_extract1$IndSigSNP) %>% filter(gwasP < 0.05) %>% 
  group_by(IndSigSNP) %>% arrange(gwasP) %>% mutate(order = row_number()) %>% 
  filter(order == 1) %>% select(-order) 
dim(final_extract2)
indsig_LBD_final = rbind(final_extract1, final_extract2) %>% unique() %>% 
  select(rsid, chr, pos, a1, a0, beta, gwasP, rsID)
dim(indsig_LBD_final) # dim = (9,8)
# Remap to get final
info_snp = snp_match(indsig_LBD_final, map_noNA) # 9 variants have been matched
indsig_beta_LBD = info_snp$beta
LBD_beta_indsig = info_snp %>% 
  select(rsid, chr, pos, a0, a1, beta, gwasP) %>% 
  dplyr::rename(beta_LBD_indsig = beta)
# Filter genotype info
genotype.targ.indsig = genotype_df_fam %>% select(all_of(LBD_beta_indsig$rsid))
genotype.targ.matrix = as.matrix(genotype.targ.indsig[])
# Calculate PRS
pred_prs = genotype.targ.matrix %*% indsig_beta_LBD
pred_prs_final = cbind(pred_prs, genotype_df_fam[, c('FID','IID')])
names(pred_prs_final) = c('LBD_PRS_indsig', 'FID', 'IID')
# Normalization
onekg_EUR = pred_prs_final %>% filter(FID == "EUR")
LBD_1kg_mean = mean(onekg_EUR$LBD_PRS_indsig) # 0.2926416
LBD_1kg_sd = sd(onekg_EUR$LBD_PRS_indsig) # 0.7824751
# Apply to ATLAS
LBD_prs_indsig = pred_prs_final %>% 
  filter(FID %!in% c("EUR", "AFR", "AMR", "EAS", "SAS"))
dim(LBD_prs_indsig) # dim = (54935,3)
LBD_prs_indsig$LBD_PRS_indsig_norm = 
  (LBD_prs_indsig$LBD_PRS_indsig - LBD_1kg_mean) / LBD_1kg_sd
LBD_prs_indsig = LBD_prs_indsig %>% 
  select(IID, LBD_PRS_indsig, LBD_PRS_indsig_norm) %>% 
  dplyr::rename(UniqueSampleId = IID)
```

### B. Lead SNPs
```{r}
final_extract = info_snp_LBD %>% filter(lead == 1) %>% unique() # N = 5
dim(final_extract)
lead_LBD_final = final_extract %>% 
  select(rsid, chr, pos, a1, a0, beta, gwasP, rsID)
dim(lead_LBD_final) # dim = (5,8)
# Remap to get final
info_snp = snp_match(lead_LBD_final, map_noNA) # 5 variants have been matched
lead_beta_LBD = info_snp$beta
LBD_beta_lead = info_snp %>% 
  select(rsid, chr, pos, a0, a1, beta, gwasP) %>% 
  dplyr::rename(beta_LBD_lead = beta)
# Filter genotype info
genotype.targ.lead = genotype_df_fam %>% select(all_of(LBD_beta_lead$rsid))
genotype.targ.matrix = as.matrix(genotype.targ.lead[])
# Calculate PRS
pred_prs = genotype.targ.matrix %*% lead_beta_LBD
pred_prs_final = cbind(pred_prs, genotype_df_fam[, c('FID','IID')])
names(pred_prs_final) = c('LBD_PRS_lead', 'FID', 'IID')
# Normalization
onekg_EUR = pred_prs_final %>% filter(FID == "EUR")
LBD_1kg_mean = mean(onekg_EUR$LBD_PRS_lead) # 0.7703105
LBD_1kg_sd = sd(onekg_EUR$LBD_PRS_lead) # 0.5765808
# Apply to ATLAS
LBD_prs_lead = pred_prs_final %>% 
  filter(FID %!in% c("EUR", "AFR", "AMR", "EAS", "SAS"))
dim(LBD_prs_lead) # dim = (54935,3)
LBD_prs_lead$LBD_PRS_lead_norm = 
  (LBD_prs_lead$LBD_PRS_lead - LBD_1kg_mean) / LBD_1kg_sd
LBD_prs_lead = LBD_prs_lead %>% 
  select(IID, LBD_PRS_lead, LBD_PRS_lead_norm) %>% 
  dplyr::rename(UniqueSampleId = IID)
```

### C. Mapped SNPs
```{r}
final_extract = info_snp_LBD %>% filter(map_yes == 1) %>% 
  group_by(IndSigSNP) %>% arrange(desc(CADD)) %>% 
  mutate(order = row_number()) %>% filter(order == 1) %>% 
  select(-order) %>% ungroup()
map_LBD_final = final_extract %>% 
  select(rsid, chr, pos, a1, a0, beta, gwasP, rsID)
dim(map_LBD_final) # dim = (9,8)
# Remap to get final
info_snp = snp_match(map_LBD_final, map_noNA) # 9 variants have been matched
map_beta_LBD = info_snp$beta
LBD_beta_map = info_snp %>% 
  select(rsid, chr, pos, a0, a1, beta, gwasP) %>% 
  dplyr::rename(beta_LBD_map = beta)
# Filter genotype info
genotype.targ.map = genotype_df_fam %>% select(all_of(LBD_beta_map$rsid))
genotype.targ.matrix = as.matrix(genotype.targ.map[])
# Calculate PRS
pred_prs = genotype.targ.matrix %*% map_beta_LBD
pred_prs_final = cbind(pred_prs, genotype_df_fam[, c('FID','IID')])
names(pred_prs_final) = c('LBD_PRS_map', 'FID', 'IID')
# Normalization
onekg_EUR = pred_prs_final %>% filter(FID == "EUR")
LBD_1kg_mean = mean(onekg_EUR$LBD_PRS_map) # -0.05188502
LBD_1kg_sd = sd(onekg_EUR$LBD_PRS_map) # 0.6901303
# Apply to ATLAS
LBD_prs_map = pred_prs_final %>% 
  filter(FID %!in% c("EUR", "AFR", "AMR", "EAS", "SAS"))
dim(LBD_prs_map) # dim = (54935,3)
LBD_prs_map$LBD_PRS_map_norm = 
  (LBD_prs_map$LBD_PRS_map - LBD_1kg_mean) / LBD_1kg_sd
LBD_prs_map = LBD_prs_map %>% 
  select(IID, LBD_PRS_map, LBD_PRS_map_norm) %>% 
  dplyr::rename(UniqueSampleId = IID)
```

### D. Mapped SNPs (independent of LD) with CADD >= 12.37
```{r}
final_extract = info_snp_LBD %>% filter(map_yes == 1) %>%
  filter(CADD >= 12.37) %>% 
  group_by(IndSigSNP) %>% arrange(desc(CADD)) %>%
  mutate(order = row_number()) %>% filter(order == 1) %>%
  select(-order) %>% ungroup()
map_CADD_LBD_final = final_extract %>% 
  select(rsid, chr, pos, a1, a0, beta, gwasP, rsID)
dim(map_CADD_LBD_final) # dim = (7,8)
# Remap to get final
info_snp = snp_match(map_CADD_LBD_final, map_noNA) # 7 variants have been matched
map_CADD_beta_LBD = info_snp$beta
LBD_beta_map_CADD = info_snp %>% 
  select(rsid, chr, pos, a0, a1, beta, gwasP) %>% 
  dplyr::rename(beta_LBD_map_CADD = beta)
# Filter genotype info
genotype.targ.map_CADD = genotype_df_fam %>% select(all_of(LBD_beta_map_CADD$rsid))
genotype.targ.matrix = as.matrix(genotype.targ.map_CADD[])
# Calculate PRS
pred_prs = genotype.targ.matrix %*% map_CADD_beta_LBD
pred_prs_final = cbind(pred_prs, genotype_df_fam[, c('FID','IID')])
names(pred_prs_final) = c('LBD_PRS_map_CADD', 'FID', 'IID')
# Normalization
onekg_EUR = pred_prs_final %>% filter(FID == "EUR")
LBD_1kg_mean = mean(onekg_EUR$LBD_PRS_map_CADD)
LBD_1kg_sd = sd(onekg_EUR$LBD_PRS_map_CADD)
# Apply to ATLAS
LBD_prs_map_CADD = pred_prs_final %>% 
  filter(FID %!in% c("EUR", "AFR", "AMR", "EAS", "SAS"))
dim(LBD_prs_map_CADD) # dim = (54935,3)
LBD_prs_map_CADD$LBD_map_CADD_norm = 
  (LBD_prs_map_CADD$LBD_PRS_map_CADD - LBD_1kg_mean) / LBD_1kg_sd
LBD_prs_map_CADD = LBD_prs_map_CADD %>% 
  select(IID, LBD_PRS_map_CADD, LBD_map_CADD_norm) %>% 
  dplyr::rename(UniqueSampleId = IID)
```

### Combine all LBD PRSs
```{r}
LBD_prs = LBD_prs_indsig %>% inner_join(LBD_prs_lead) %>% 
  inner_join(LBD_prs_map) %>% inner_join(LBD_prs_map_CADD)
LBD_beta = LBD_beta_indsig %>% full_join(LBD_beta_lead) %>% 
  full_join(LBD_beta_map) %>% full_join(LBD_beta_map_CADD) %>% select(-gwasP)
save(LBD_prs, file = paste0(raw_data_path, "prs/raw/LBD_prs.rda"))
save(LBD_beta, file = paste0(raw_data_path, "prs/raw/LBD_beta.rda"))
```

## 5. PD
```{r}
snp_PD = overlap_fuma %>% filter(phenotype == "PD") %>% 
  mutate(ind_sig = if_else(rsID == IndSigSNP, 1, 0)) %>% filter(!is.na(beta)) %>% 
  mutate(map_yes = case_when(
    posMapFilt == 1 | eqtlMapFilt == 1 | ciMapFilt == 1 ~ 1,
    TRUE ~ 0
  )) %>% select(chr_num, pos, a1, a0, beta, gwasP, rsID, IndSigSNP, 
                ind_sig, lead, map_yes, CADD) %>% unique() %>% 
  dplyr::rename(chr = chr_num)
# Reverse beta if necessary by performing SNP matching
info_snp_PD = snp_match(snp_PD, map_noNA) # 2640 variants have been matched
```

### A. Independent significant SNPs 
```{r}
final_extract1 = info_snp_PD %>% filter(ind_sig == 1) %>% unique() # N = 11
dim(final_extract1)
final_extract2 = info_snp_PD %>% filter(ind_sig == 0) %>% 
  filter(IndSigSNP %!in% final_extract1$IndSigSNP) %>% filter(gwasP < 0.05) %>% 
  group_by(IndSigSNP) %>% arrange(gwasP) %>% mutate(order = row_number()) %>% 
  filter(order == 1) %>% select(-order) 
dim(final_extract2)
indsig_PD_final = rbind(final_extract1, final_extract2) %>% unique() %>% 
  select(rsid, chr, pos, a1, a0, beta, gwasP, rsID)
dim(indsig_PD_final) # dim = (27,8)
# Remap to get final
info_snp = snp_match(indsig_PD_final, map_noNA) # 27 variants have been matched
indsig_beta_PD = info_snp$beta
PD_beta_indsig = info_snp %>% 
  select(rsid, chr, pos, a0, a1, beta, gwasP) %>% 
  dplyr::rename(beta_PD_indsig = beta)
# Filter genotype info
genotype.targ.indsig = genotype_df_fam %>% select(all_of(PD_beta_indsig$rsid))
genotype.targ.matrix = as.matrix(genotype.targ.indsig[])
# Calculate PRS
pred_prs = genotype.targ.matrix %*% indsig_beta_PD
pred_prs_final = cbind(pred_prs, genotype_df_fam[, c('FID','IID')])
names(pred_prs_final) = c('PD_PRS_indsig', 'FID', 'IID')
# Normalization
onekg_EUR = pred_prs_final %>% filter(FID == "EUR")
PD_1kg_mean = mean(onekg_EUR$PD_PRS_indsig) # -0.4156226
PD_1kg_sd = sd(onekg_EUR$PD_PRS_indsig) # 0.5244261
# Apply to ATLAS
PD_prs_indsig = pred_prs_final %>% 
  filter(FID %!in% c("EUR", "AFR", "AMR", "EAS", "SAS"))
dim(PD_prs_indsig) # dim = (54935,3)
PD_prs_indsig$PD_PRS_indsig_norm = 
  (PD_prs_indsig$PD_PRS_indsig - PD_1kg_mean) / PD_1kg_sd
PD_prs_indsig = PD_prs_indsig %>% 
  select(IID, PD_PRS_indsig, PD_PRS_indsig_norm) %>% 
  dplyr::rename(UniqueSampleId = IID)
```

### B. Lead SNPs
```{r}
final_extract = info_snp_PD %>% filter(lead == 1) %>% unique() # N = 9
dim(final_extract)
lead_PD_final = final_extract %>% 
  select(rsid, chr, pos, a1, a0, beta, gwasP, rsID)
dim(lead_PD_final) # dim = (9,8)
# Remap to get final
info_snp = snp_match(lead_PD_final, map_noNA) # 9 variants have been matched
lead_beta_PD = info_snp$beta
PD_beta_lead = info_snp %>% 
  select(rsid, chr, pos, a0, a1, beta, gwasP) %>% 
  dplyr::rename(beta_PD_lead = beta)
# Filter genotype info
genotype.targ.lead = genotype_df_fam %>% select(all_of(PD_beta_lead$rsid))
genotype.targ.matrix = as.matrix(genotype.targ.lead[])
# Calculate PRS
pred_prs = genotype.targ.matrix %*% lead_beta_PD
pred_prs_final = cbind(pred_prs, genotype_df_fam[, c('FID','IID')])
names(pred_prs_final) = c('PD_PRS_lead', 'FID', 'IID')
# Normalization
onekg_EUR = pred_prs_final %>% filter(FID == "EUR")
PD_1kg_mean = mean(onekg_EUR$PD_PRS_lead) # -0.07185268
PD_1kg_sd = sd(onekg_EUR$PD_PRS_lead) # 0.2589001
# Apply to ATLAS
PD_prs_lead = pred_prs_final %>% 
  filter(FID %!in% c("EUR", "AFR", "AMR", "EAS", "SAS"))
dim(PD_prs_lead) # dim = (54935,3)
PD_prs_lead$PD_PRS_lead_norm = 
  (PD_prs_lead$PD_PRS_lead - PD_1kg_mean) / PD_1kg_sd
PD_prs_lead = PD_prs_lead %>% 
  select(IID, PD_PRS_lead, PD_PRS_lead_norm) %>% 
  dplyr::rename(UniqueSampleId = IID)
```

### C. Mapped SNPs
```{r}
final_extract = info_snp_PD %>% filter(map_yes == 1) %>% 
  group_by(IndSigSNP) %>% arrange(desc(CADD)) %>% 
  mutate(order = row_number()) %>% filter(order == 1) %>% 
  select(-order) %>% ungroup()
map_PD_final = final_extract %>% 
  select(rsid, chr, pos, a1, a0, beta, gwasP, rsID)
dim(map_PD_final) # dim = (27,8)
# Remap to get final
info_snp = snp_match(map_PD_final, map_noNA) # 27 variants have been matched
map_beta_PD = info_snp$beta
PD_beta_map = info_snp %>% 
  select(rsid, chr, pos, a0, a1, beta, gwasP) %>% 
  dplyr::rename(beta_PD_map = beta)
# Filter genotype info
genotype.targ.map = genotype_df_fam %>% select(all_of(PD_beta_map$rsid))
genotype.targ.matrix = as.matrix(genotype.targ.map[])
# Calculate PRS
pred_prs = genotype.targ.matrix %*% map_beta_PD
pred_prs_final = cbind(pred_prs, genotype_df_fam[, c('FID','IID')])
names(pred_prs_final) = c('PD_PRS_map', 'FID', 'IID')
# Normalization
onekg_EUR = pred_prs_final %>% filter(FID == "EUR")
PD_1kg_mean = mean(onekg_EUR$PD_PRS_map) # -0.2469264
PD_1kg_sd = sd(onekg_EUR$PD_PRS_map) # 0.3098816
# Apply to ATLAS
PD_prs_map = pred_prs_final %>% 
  filter(FID %!in% c("EUR", "AFR", "AMR", "EAS", "SAS"))
dim(PD_prs_map) # dim = (54935,3)
PD_prs_map$PD_PRS_map_norm = 
  (PD_prs_map$PD_PRS_map - PD_1kg_mean) / PD_1kg_sd
PD_prs_map = PD_prs_map %>% 
  select(IID, PD_PRS_map, PD_PRS_map_norm) %>% 
  dplyr::rename(UniqueSampleId = IID)
```

### D. Mapped SNPs (independent of LD) with CADD >= 12.37
```{r}
final_extract = info_snp_PD %>% filter(map_yes == 1) %>%
  filter(CADD >= 12.37) %>% 
  group_by(IndSigSNP) %>% arrange(desc(CADD)) %>%
  mutate(order = row_number()) %>% filter(order == 1) %>%
  select(-order) %>% ungroup()
map_CADD_PD_final = final_extract %>% 
  select(rsid, chr, pos, a1, a0, beta, gwasP, rsID)
dim(map_CADD_PD_final) # dim = (23,8)
# Remap to get final
info_snp = snp_match(map_CADD_PD_final, map_noNA) # 23 variants have been matched
map_CADD_beta_PD = info_snp$beta
PD_beta_map_CADD = info_snp %>% 
  select(rsid, chr, pos, a0, a1, beta, gwasP) %>% 
  dplyr::rename(beta_PD_map_CADD = beta)
# Filter genotype info
genotype.targ.map_CADD = genotype_df_fam %>% select(all_of(PD_beta_map_CADD$rsid))
genotype.targ.matrix = as.matrix(genotype.targ.map_CADD[])
# Calculate PRS
pred_prs = genotype.targ.matrix %*% map_CADD_beta_PD
pred_prs_final = cbind(pred_prs, genotype_df_fam[, c('FID','IID')])
names(pred_prs_final) = c('PD_PRS_map_CADD', 'FID', 'IID')
# Normalization
onekg_EUR = pred_prs_final %>% filter(FID == "EUR")
PD_1kg_mean = mean(onekg_EUR$PD_PRS_map_CADD)
PD_1kg_sd = sd(onekg_EUR$PD_PRS_map_CADD)
# Apply to ATLAS
PD_prs_map_CADD = pred_prs_final %>% 
  filter(FID %!in% c("EUR", "AFR", "AMR", "EAS", "SAS"))
dim(PD_prs_map_CADD) # dim = (54935,3)
PD_prs_map_CADD$PD_map_CADD_norm = 
  (PD_prs_map_CADD$PD_PRS_map_CADD - PD_1kg_mean) / PD_1kg_sd
PD_prs_map_CADD = PD_prs_map_CADD %>% 
  select(IID, PD_PRS_map_CADD, PD_map_CADD_norm) %>% 
  dplyr::rename(UniqueSampleId = IID)
```

### Combine all PD PRSs
```{r}
PD_prs = PD_prs_indsig %>% inner_join(PD_prs_lead) %>% 
  inner_join(PD_prs_map) %>% inner_join(PD_prs_map_CADD)
PD_beta = PD_beta_indsig %>% full_join(PD_beta_lead) %>% 
  full_join(PD_beta_map) %>% full_join(PD_beta_map_CADD) %>% select(-gwasP)
save(PD_prs, file = paste0(raw_data_path, "prs/raw/PD_prs.rda"))
save(PD_beta, file = paste0(raw_data_path, "prs/raw/PD_beta.rda"))
```

## 6. PSP
```{r}
snp_PSP = overlap_fuma %>% filter(phenotype == "PSP") %>% 
  mutate(ind_sig = if_else(rsID == IndSigSNP, 1, 0)) %>% filter(!is.na(beta)) %>% 
  mutate(map_yes = case_when(
    posMapFilt == 1 | eqtlMapFilt == 1 | ciMapFilt == 1 ~ 1,
    TRUE ~ 0
  )) %>% select(chr_num, pos, a1, a0, beta, gwasP, rsID, IndSigSNP, 
                ind_sig, lead, map_yes, CADD) %>% unique() %>% 
  dplyr::rename(chr = chr_num)
# Reverse beta if necessary by performing SNP matching
info_snp_PSP = snp_match(snp_PSP, map_noNA, match.min.prop = 0) # 732 variants have been matched
```

### A. Independent significant SNPs 
```{r}
final_extract1 = info_snp_PSP %>% filter(ind_sig == 1) %>% unique() # N = 8
dim(final_extract1)
final_extract2 = info_snp_PSP %>% filter(ind_sig == 0) %>% 
  filter(IndSigSNP %!in% final_extract1$IndSigSNP) %>% filter(gwasP < 0.05) %>% 
  group_by(IndSigSNP) %>% arrange(gwasP) %>% mutate(order = row_number()) %>% 
  filter(order == 1) %>% select(-order) 
dim(final_extract2)
indsig_PSP_final = rbind(final_extract1, final_extract2) %>% unique() %>% 
  select(rsid, chr, pos, a1, a0, beta, gwasP, rsID)
dim(indsig_PSP_final) # dim = (21,8)
# Remap to get final
info_snp = snp_match(indsig_PSP_final, map_noNA) # 21 variants have been matched
indsig_beta_PSP = info_snp$beta
PSP_beta_indsig = info_snp %>% 
  select(rsid, chr, pos, a0, a1, beta, gwasP) %>% 
  dplyr::rename(beta_PSP_indsig = beta)
# Filter genotype info
genotype.targ.indsig = genotype_df_fam %>% select(all_of(PSP_beta_indsig$rsid))
genotype.targ.matrix = as.matrix(genotype.targ.indsig[])
# Calculate PRS
pred_prs = genotype.targ.matrix %*% indsig_beta_PSP
pred_prs_final = cbind(pred_prs, genotype_df_fam[, c('FID','IID')])
names(pred_prs_final) = c('PSP_PRS_indsig', 'FID', 'IID')
# Normalization
onekg_EUR = pred_prs_final %>% filter(FID == "EUR")
PSP_1kg_mean = mean(onekg_EUR$PSP_PRS_indsig) # 0.2032141
PSP_1kg_sd = sd(onekg_EUR$PSP_PRS_indsig) # 0.1415184
# Apply to ATLAS
PSP_prs_indsig = pred_prs_final %>% 
  filter(FID %!in% c("EUR", "AFR", "AMR", "EAS", "SAS"))
dim(PSP_prs_indsig) # dim = (54935,3)
PSP_prs_indsig$PSP_PRS_indsig_norm = 
  (PSP_prs_indsig$PSP_PRS_indsig - PSP_1kg_mean) / PSP_1kg_sd
PSP_prs_indsig = PSP_prs_indsig %>% 
  select(IID, PSP_PRS_indsig, PSP_PRS_indsig_norm) %>% 
  dplyr::rename(UniqueSampleId = IID)
```

### B. Lead SNPs
```{r}
final_extract = info_snp_PSP %>% filter(lead == 1) %>% unique() # N = 9
lead_PSP_final = final_extract %>% 
  select(rsid, chr, pos, a1, a0, beta, gwasP, rsID)
dim(lead_PSP_final) # dim = (9,8)
# Remap to get final
info_snp = snp_match(lead_PSP_final, map_noNA) # 9 variants have been matched
lead_beta_PSP = info_snp$beta
PSP_beta_lead = info_snp %>% 
  select(rsid, chr, pos, a0, a1, beta, gwasP) %>% 
  dplyr::rename(beta_PSP_lead = beta)
# Filter genotype info
genotype.targ.lead = genotype_df_fam %>% select(all_of(PSP_beta_lead$rsid))
genotype.targ.matrix = as.matrix(genotype.targ.lead[])
# Calculate PRS
pred_prs = genotype.targ.matrix %*% lead_beta_PSP
pred_prs_final = cbind(pred_prs, genotype_df_fam[, c('FID','IID')])
names(pred_prs_final) = c('PSP_PRS_lead', 'FID', 'IID')
# Normalization
onekg_EUR = pred_prs_final %>% filter(FID == "EUR")
PSP_1kg_mean = mean(onekg_EUR$PSP_PRS_lead) # -0.03481778
PSP_1kg_sd = sd(onekg_EUR$PSP_PRS_lead) # 0.09599806
# Apply to ATLAS
PSP_prs_lead = pred_prs_final %>% 
  filter(FID %!in% c("EUR", "AFR", "AMR", "EAS", "SAS"))
dim(PSP_prs_lead) # dim = (54935,3)
PSP_prs_lead$PSP_PRS_lead_norm = 
  (PSP_prs_lead$PSP_PRS_lead - PSP_1kg_mean) / PSP_1kg_sd
PSP_prs_lead = PSP_prs_lead %>% 
  select(IID, PSP_PRS_lead, PSP_PRS_lead_norm) %>% 
  dplyr::rename(UniqueSampleId = IID)
```

### C. Mapped SNPs
```{r}
final_extract = info_snp_PSP %>% filter(map_yes == 1) %>% 
  group_by(IndSigSNP) %>% arrange(desc(CADD)) %>% 
  mutate(order = row_number()) %>% filter(order == 1) %>% 
  select(-order) %>% ungroup()
map_PSP_final = final_extract %>% 
  select(rsid, chr, pos, a1, a0, beta, gwasP, rsID)
dim(map_PSP_final) # dim = (21,8)
# Remap to get final
info_snp = snp_match(map_PSP_final, map_noNA) # 21 variants have been matched
map_beta_PSP = info_snp$beta
PSP_beta_map = info_snp %>% 
  select(rsid, chr, pos, a0, a1, beta, gwasP) %>% 
  dplyr::rename(beta_PSP_map = beta)
# Filter genotype info
genotype.targ.map = genotype_df_fam %>% select(all_of(PSP_beta_map$rsid))
genotype.targ.matrix = as.matrix(genotype.targ.map[])
# Calculate PRS
pred_prs = genotype.targ.matrix %*% map_beta_PSP
pred_prs_final = cbind(pred_prs, genotype_df_fam[, c('FID','IID')])
names(pred_prs_final) = c('PSP_PRS_map', 'FID', 'IID')
# Normalization
onekg_EUR = pred_prs_final %>% filter(FID == "EUR")
PSP_1kg_mean = mean(onekg_EUR$PSP_PRS_map) # 0.07864808
PSP_1kg_sd = sd(onekg_EUR$PSP_PRS_map) # 0.1252794
# Apply to ATLAS
PSP_prs_map = pred_prs_final %>% 
  filter(FID %!in% c("EUR", "AFR", "AMR", "EAS", "SAS"))
dim(PSP_prs_map) # dim = (54935,3)
PSP_prs_map$PSP_PRS_map_norm = 
  (PSP_prs_map$PSP_PRS_map - PSP_1kg_mean) / PSP_1kg_sd
PSP_prs_map = PSP_prs_map %>% 
  select(IID, PSP_PRS_map, PSP_PRS_map_norm) %>% 
  dplyr::rename(UniqueSampleId = IID)
```

### D. Mapped SNPs (independent of LD) with CADD >= 12.37
```{r}
final_extract = info_snp_PSP %>% filter(map_yes == 1) %>%
  filter(CADD >= 12.37) %>% 
  group_by(IndSigSNP) %>% arrange(desc(CADD)) %>%
  mutate(order = row_number()) %>% filter(order == 1) %>%
  select(-order) %>% ungroup()
map_CADD_PSP_final = final_extract %>% 
  select(rsid, chr, pos, a1, a0, beta, gwasP, rsID)
dim(map_CADD_PSP_final) # dim = (7,8)
# Remap to get final
info_snp = snp_match(map_CADD_PSP_final, map_noNA) # 7 variants have been matched
map_CADD_beta_PSP = info_snp$beta
PSP_beta_map_CADD = info_snp %>% 
  select(rsid, chr, pos, a0, a1, beta, gwasP) %>% 
  dplyr::rename(beta_PSP_map_CADD = beta)
# Filter genotype info
genotype.targ.map_CADD = genotype_df_fam %>% select(all_of(PSP_beta_map_CADD$rsid))
genotype.targ.matrix = as.matrix(genotype.targ.map_CADD[])
# Calculate PRS
pred_prs = genotype.targ.matrix %*% map_CADD_beta_PSP
pred_prs_final = cbind(pred_prs, genotype_df_fam[, c('FID','IID')])
names(pred_prs_final) = c('PSP_PRS_map_CADD', 'FID', 'IID')
# Normalization
onekg_EUR = pred_prs_final %>% filter(FID == "EUR")
PSP_1kg_mean = mean(onekg_EUR$PSP_PRS_map_CADD)
PSP_1kg_sd = sd(onekg_EUR$PSP_PRS_map_CADD)
# Apply to ATLAS
PSP_prs_map_CADD = pred_prs_final %>% 
  filter(FID %!in% c("EUR", "AFR", "AMR", "EAS", "SAS"))
dim(PSP_prs_map_CADD) # dim = (54935,3)
PSP_prs_map_CADD$PSP_map_CADD_norm = 
  (PSP_prs_map_CADD$PSP_PRS_map_CADD - PSP_1kg_mean) / PSP_1kg_sd
PSP_prs_map_CADD = PSP_prs_map_CADD %>% 
  select(IID, PSP_PRS_map_CADD, PSP_map_CADD_norm) %>% 
  dplyr::rename(UniqueSampleId = IID)
```

### Combine all PSP PRSs
```{r}
PSP_prs = PSP_prs_indsig %>% inner_join(PSP_prs_lead) %>% 
  inner_join(PSP_prs_map) %>% inner_join(PSP_prs_map_CADD)
PSP_beta = PSP_beta_indsig %>% full_join(PSP_beta_lead) %>% 
  full_join(PSP_beta_map) %>% full_join(PSP_beta_map_CADD) %>% select(-gwasP)
save(PSP_prs, file = paste0(raw_data_path, "prs/raw/PSP_prs.rda"))
save(PSP_beta, file = paste0(raw_data_path, "prs/raw/PSP_beta.rda"))
```

## 7. STROKE
```{r}
snp_STROKE = overlap_fuma %>% filter(phenotype == "STROKE") %>% 
  mutate(ind_sig = if_else(rsID == IndSigSNP, 1, 0)) %>% filter(!is.na(beta)) %>% 
  mutate(map_yes = case_when(
    posMapFilt == 1 | eqtlMapFilt == 1 | ciMapFilt == 1 ~ 1,
    TRUE ~ 0
  )) %>% select(chr_num, pos, a1, a0, beta, gwasP, rsID, IndSigSNP, 
                ind_sig, lead, map_yes, CADD) %>% unique() %>% 
  dplyr::rename(chr = chr_num)
# Reverse beta if necessary by performing SNP matching
info_snp_STROKE = snp_match(snp_STROKE, map_noNA) # 903 variants have been matched
```

### A. Independent significant SNPs 
```{r}
final_extract1 = info_snp_STROKE %>% filter(ind_sig == 1) %>% unique() # N = 13
dim(final_extract1)
final_extract2 = info_snp_STROKE %>% filter(ind_sig == 0) %>% 
  filter(IndSigSNP %!in% final_extract1$IndSigSNP) %>% filter(gwasP < 0.05) %>% 
  group_by(IndSigSNP) %>% arrange(gwasP) %>% mutate(order = row_number()) %>% 
  filter(order == 1) %>% select(-order)
dim(final_extract2)
indsig_STROKE_final = rbind(final_extract1, final_extract2) %>% unique() %>% 
  select(rsid, chr, pos, a1, a0, beta, gwasP, rsID)
dim(indsig_STROKE_final) # dim = (21,8)
# Remap to get final
info_snp = snp_match(indsig_STROKE_final, map_noNA) # 21 variants have been matched
indsig_beta_STROKE = info_snp$beta
STROKE_beta_indsig = info_snp %>% 
  select(rsid, chr, pos, a0, a1, beta, gwasP) %>% 
  dplyr::rename(beta_STROKE_indsig = beta)
# Filter genotype info
genotype.targ.indsig = genotype_df_fam %>% select(all_of(STROKE_beta_indsig$rsid))
genotype.targ.matrix = as.matrix(genotype.targ.indsig[])
# Calculate PRS
pred_prs = genotype.targ.matrix %*% indsig_beta_STROKE
pred_prs_final = cbind(pred_prs, genotype_df_fam[, c('FID','IID')])
names(pred_prs_final) = c('STROKE_PRS_indsig', 'FID', 'IID')
# Normalization
onekg_EUR = pred_prs_final %>% filter(FID == "EUR")
STROKE_1kg_mean = mean(onekg_EUR$STROKE_PRS_indsig) # 0.169449
STROKE_1kg_sd = sd(onekg_EUR$STROKE_PRS_indsig) # 0.1939814
# Apply to ATLAS
STROKE_prs_indsig = pred_prs_final %>% 
  filter(FID %!in% c("EUR", "AFR", "AMR", "EAS", "SAS"))
dim(STROKE_prs_indsig) # dim = (54935,3)
STROKE_prs_indsig$STROKE_PRS_indsig_norm = 
  (STROKE_prs_indsig$STROKE_PRS_indsig - STROKE_1kg_mean) / STROKE_1kg_sd
STROKE_prs_indsig = STROKE_prs_indsig %>% 
  select(IID, STROKE_PRS_indsig, STROKE_PRS_indsig_norm) %>% 
  dplyr::rename(UniqueSampleId = IID)
```

### B. Lead SNPs
```{r}
final_extract = info_snp_STROKE %>% filter(lead == 1) %>% unique() # N = 11
lead_STROKE_final = final_extract %>% 
  select(rsid, chr, pos, a1, a0, beta, gwasP, rsID)
dim(lead_STROKE_final) # dim = (11,8)
# Remap to get final
info_snp = snp_match(lead_STROKE_final, map_noNA) # 11 variants have been matched
lead_beta_STROKE = info_snp$beta
STROKE_beta_lead = info_snp %>% 
  select(rsid, chr, pos, a0, a1, beta, gwasP) %>% 
  dplyr::rename(beta_STROKE_lead = beta)
# Filter genotype info
genotype.targ.lead = genotype_df_fam %>% select(all_of(STROKE_beta_lead$rsid))
genotype.targ.matrix = as.matrix(genotype.targ.lead[])
# Calculate PRS
pred_prs = genotype.targ.matrix %*% lead_beta_STROKE
pred_prs_final = cbind(pred_prs, genotype_df_fam[, c('FID','IID')])
names(pred_prs_final) = c('STROKE_PRS_lead', 'FID', 'IID')
# Normalization
onekg_EUR = pred_prs_final %>% filter(FID == "EUR")
STROKE_1kg_mean = mean(onekg_EUR$STROKE_PRS_lead) # -0.03166743
STROKE_1kg_sd = sd(onekg_EUR$STROKE_PRS_lead) # 0.1237029
# Apply to ATLAS
STROKE_prs_lead = pred_prs_final %>% 
  filter(FID %!in% c("EUR", "AFR", "AMR", "EAS", "SAS"))
dim(STROKE_prs_lead) # dim = (54935,3)
STROKE_prs_lead$STROKE_PRS_lead_norm = 
  (STROKE_prs_lead$STROKE_PRS_lead - STROKE_1kg_mean) / STROKE_1kg_sd
STROKE_prs_lead = STROKE_prs_lead %>% 
  select(IID, STROKE_PRS_lead, STROKE_PRS_lead_norm) %>% 
  dplyr::rename(UniqueSampleId = IID)
```

### C. Mapped SNPs
```{r}
final_extract = info_snp_STROKE %>% filter(map_yes == 1) %>% 
  group_by(IndSigSNP) %>% arrange(desc(CADD)) %>% 
  mutate(order = row_number()) %>% filter(order == 1) %>% 
  select(-order) %>% ungroup()
map_STROKE_final = final_extract %>% 
  select(rsid, chr, pos, a1, a0, beta, gwasP, rsID)
dim(map_STROKE_final) # dim = (20,8)
# Remap to get final
info_snp = snp_match(map_STROKE_final, map_noNA) # 20 variants have been matched
map_beta_STROKE = info_snp$beta
STROKE_beta_map = info_snp %>% 
  select(rsid, chr, pos, a0, a1, beta, gwasP) %>% 
  dplyr::rename(beta_STROKE_map = beta)
# Filter genotype info
genotype.targ.map = genotype_df_fam %>% select(all_of(STROKE_beta_map$rsid))
genotype.targ.matrix = as.matrix(genotype.targ.map[])
# Calculate PRS
pred_prs = genotype.targ.matrix %*% map_beta_STROKE
pred_prs_final = cbind(pred_prs, genotype_df_fam[, c('FID','IID')])
names(pred_prs_final) = c('STROKE_PRS_map', 'FID', 'IID')
# Normalization
onekg_EUR = pred_prs_final %>% filter(FID == "EUR")
STROKE_1kg_mean = mean(onekg_EUR$STROKE_PRS_map) # 0.009353257
STROKE_1kg_sd = sd(onekg_EUR$STROKE_PRS_map) # 0.1423629
# Apply to ATLAS
STROKE_prs_map = pred_prs_final %>% 
  filter(FID %!in% c("EUR", "AFR", "AMR", "EAS", "SAS"))
dim(STROKE_prs_map) # dim = (54935,3)
STROKE_prs_map$STROKE_PRS_map_norm = 
  (STROKE_prs_map$STROKE_PRS_map - STROKE_1kg_mean) / STROKE_1kg_sd
STROKE_prs_map = STROKE_prs_map %>% 
  select(IID, STROKE_PRS_map, STROKE_PRS_map_norm) %>% 
  dplyr::rename(UniqueSampleId = IID)
```

### D. Mapped SNPs (independent of LD) with CADD >= 12.37
```{r}
final_extract = info_snp_STROKE %>% filter(map_yes == 1) %>%
  filter(CADD >= 12.37) %>% 
  group_by(IndSigSNP) %>% arrange(desc(CADD)) %>%
  mutate(order = row_number()) %>% filter(order == 1) %>%
  select(-order) %>% ungroup()
map_CADD_STROKE_final = final_extract %>% 
  select(rsid, chr, pos, a1, a0, beta, gwasP, rsID)
dim(map_CADD_STROKE_final) # dim = (15,8)
# Remap to get final
info_snp = snp_match(map_CADD_STROKE_final, map_noNA) # 15 variants have been matched
map_CADD_beta_STROKE = info_snp$beta
STROKE_beta_map_CADD = info_snp %>% 
  select(rsid, chr, pos, a0, a1, beta, gwasP) %>% 
  dplyr::rename(beta_STROKE_map_CADD = beta)
# Filter genotype info
genotype.targ.map_CADD = genotype_df_fam %>% select(all_of(STROKE_beta_map_CADD$rsid))
genotype.targ.matrix = as.matrix(genotype.targ.map_CADD[])
# Calculate PRS
pred_prs = genotype.targ.matrix %*% map_CADD_beta_STROKE
pred_prs_final = cbind(pred_prs, genotype_df_fam[, c('FID','IID')])
names(pred_prs_final) = c('STROKE_PRS_map_CADD', 'FID', 'IID')
# Normalization
onekg_EUR = pred_prs_final %>% filter(FID == "EUR")
STROKE_1kg_mean = mean(onekg_EUR$STROKE_PRS_map_CADD)
STROKE_1kg_sd = sd(onekg_EUR$STROKE_PRS_map_CADD)
# Apply to ATLAS
STROKE_prs_map_CADD = pred_prs_final %>% 
  filter(FID %!in% c("EUR", "AFR", "AMR", "EAS", "SAS"))
dim(STROKE_prs_map_CADD) # dim = (54935,3)
STROKE_prs_map_CADD$STROKE_map_CADD_norm = 
  (STROKE_prs_map_CADD$STROKE_PRS_map_CADD - STROKE_1kg_mean) / STROKE_1kg_sd
STROKE_prs_map_CADD = STROKE_prs_map_CADD %>% 
  select(IID, STROKE_PRS_map_CADD, STROKE_map_CADD_norm) %>% 
  dplyr::rename(UniqueSampleId = IID)
```

### Combine all STROKE PRSs
```{r}
STROKE_prs = STROKE_prs_indsig %>% inner_join(STROKE_prs_lead) %>% 
  inner_join(STROKE_prs_map) %>% inner_join(STROKE_prs_map_CADD)
STROKE_beta = STROKE_beta_indsig %>% full_join(STROKE_beta_lead) %>% 
  full_join(STROKE_beta_map) %>% full_join(STROKE_beta_map_CADD) %>% select(-gwasP)
save(STROKE_prs, file = paste0(raw_data_path, "prs/raw/STROKE_prs.rda"))
save(STROKE_beta, file = paste0(raw_data_path, "prs/raw/STROKE_beta.rda"))
```


# Part 4. Combine PRSs from all GWASs
```{r}
rm(list = ls())
raw_data_path = "/Users/Mingzhou/Desktop/Projects/Dementia-prediction/data/"
output_path = "/Users/Mingzhou/Desktop/Projects/Dementia-prediction/output/"
# Source in useful functions
source("/Users/Mingzhou/Desktop/Projects/Dementia-prediction/code/functions.R")
# Read in PRS + beta data
prs_folder = paste0(raw_data_path, "prs/raw") 
prs_files = list.files(prs_folder, pattern = ".rda$")
for (i in prs_files) {
  load(paste0(raw_data_path, 'prs/raw/', i))
}
```

## 1. Merge all PRSs
```{r}
PRS_merge_raw = AD_Kunkle_EUR_prs %>% full_join(AD_Kunkle_AFR_prs) %>% 
  full_join(AD_Jun_trans_prs) %>% full_join(LBD_prs) %>% 
  full_join(PD_prs) %>% full_join(PSP_prs) %>% full_join(STROKE_prs) %>% 
  select(UniqueSampleId, AD_PRS_EUR_indsig_norm, AD_PRS_EUR_lead_norm, 
         AD_PRS_EUR_map_norm, AD_PRS_EUR_map_CADD_norm,
         AD_PRS_AFR_indsig_norm, AD_PRS_AFR_lead_norm, AD_PRS_AFR_map_norm,
         AD_PRS_AFR_map_CADD_norm, AD_PRS_trans_indsig_norm, 
         AD_PRS_trans_lead_norm, AD_PRS_trans_map_norm, 
         AD_PRS_trans_map_CADD_norm, LBD_PRS_indsig_norm, 
         LBD_PRS_lead_norm, LBD_PRS_map_norm, LBD_map_CADD_norm,
         PD_PRS_indsig_norm, PD_PRS_lead_norm, PD_PRS_map_norm,
         PD_map_CADD_norm, PSP_PRS_indsig_norm, PSP_PRS_lead_norm, 
         PSP_PRS_map_norm, PSP_map_CADD_norm, STROKE_PRS_indsig_norm, 
         STROKE_PRS_lead_norm, STROKE_PRS_map_norm, STROKE_map_CADD_norm)
# Load in ancestry info + apoe
load(file = paste0(raw_data_path, "atlas/geno/apoe.rda"))
load(file = paste0(raw_data_path, "atlas/geno/gen_ancestry.rda"))
atlas_prs_final = gen_ancestry %>% 
  mutate(UniqueSampleId = as.character(UniqueSampleId)) %>% 
  inner_join(PRS_merge_raw) %>% inner_join(apoe) 
names(atlas_prs_final) = c("UniqueSampleId", "gen_ancestry", 
                           "AD_PRS_EUR_indsig", "AD_PRS_EUR_lead", 
                           "AD_PRS_EUR_map",  "AD_PRS_EUR_map_CADD",  
                           "AD_PRS_AFR_indsig", "AD_PRS_AFR_lead", 
                           "AD_PRS_AFR_map", "AD_PRS_AFR_map_CADD",
                           "AD_PRS_Trans_indsig", "AD_PRS_Trans_lead", 
                           "AD_PRS_Trans_map", "AD_PRS_Trans_map_CADD", 
                           "LBD_PRS_indsig", "LBD_PRS_lead", 
                           "LBD_PRS_map", "LBD_PRS_map_CADD",
                           "PD_PRS_indsig", "PD_PRS_lead", 
                           "PD_PRS_map", "PD_PRS_map_CADD",
                           "PSP_PRS_indsig", "PSP_PRS_lead", 
                           "PSP_PRS_map", "PSP_PRS_map_CADD",
                           "Stroke_PRS_indsig", "Stroke_PRS_lead", 
                           "Stroke_PRS_map", "Stroke_PRS_map_CADD",
                           "APOE", "e4count", "e2count")
dim(atlas_prs_final) # dim = c(51246,33)
save(atlas_prs_final, file = paste0(raw_data_path, "prs/mod/atlas_prs_final.rda"))
```

## 2. Combine all beta info from GWASs
```{r}
beta_merge_raw = AD_Kunkle_EUR_beta %>% 
  full_join(AD_Kunkle_AFR_beta, by = c("rsid", "chr", "pos", "a0", "a1")) %>% 
  full_join(AD_Jun_trans_beta, by = c("rsid", "chr", "pos", "a0", "a1")) %>% 
  full_join(LBD_beta, by = c("rsid", "chr", "pos", "a0", "a1")) %>% 
  full_join(PD_beta, by = c("rsid", "chr", "pos", "a0", "a1")) %>% 
  full_join(PSP_beta, by = c("rsid", "chr", "pos", "a0", "a1")) %>% 
  full_join(STROKE_beta, by = c("rsid", "chr", "pos", "a0", "a1")) 
dim(beta_merge_raw) # dim = (327,33)
save(beta_merge_raw, file = paste0(raw_data_path, "modeling/beta_merge_raw.rda"))
```

## 3. Extract combined SNPs
```{r}
# AD ind.sig SNPs
AD.indsig.combine = beta_merge_raw %>% 
  filter(!is.na(beta_EUR_indsig) | !is.na(beta_AFR_indsig) | !is.na(beta_trans_indsig)) %>% 
  rowwise() %>% 
  mutate(beta_AD_indsig = mean(c(beta_EUR_indsig, beta_AFR_indsig, beta_trans_indsig), na.rm = T)) %>% 
  select(rsid, chr, pos, a0, a1, beta_EUR_indsig, beta_AFR_indsig, 
         beta_trans_indsig, beta_AD_indsig)
AD.indsig.combine.id = AD.indsig.combine %>% pull(rsid) %>% unique()
length(AD.indsig.combine.id) # 118
save(AD.indsig.combine.id, file = paste0(raw_data_path, "modeling/snp_ids/AD.indsig.combine.id.rda"))
# AD lead SNPs
AD.lead.combine = beta_merge_raw %>% 
  filter(!is.na(beta_EUR_lead) | !is.na(beta_AFR_lead) | !is.na(beta_trans_lead)) %>% 
  rowwise() %>% 
  mutate(beta_AD_lead = mean(c(beta_EUR_lead, beta_AFR_lead, beta_trans_lead), na.rm = T)) %>% 
  select(rsid, chr, pos, a0, a1, beta_EUR_lead, beta_AFR_lead, 
         beta_trans_lead, beta_AD_lead)
AD.lead.combine.id = AD.lead.combine %>% pull(rsid) %>% unique()
length(AD.lead.combine.id) # 45
save(AD.lead.combine.id, file = paste0(raw_data_path, "modeling/snp_ids/AD.lead.combine.id.rda"))
# AD mapped SNPs
AD.map.combine = beta_merge_raw %>% 
  filter(!is.na(beta_EUR_map) | !is.na(beta_AFR_map) | !is.na(beta_trans_map)) %>% 
  rowwise() %>% 
  mutate(beta_AD_map = mean(c(beta_EUR_map, beta_AFR_map, beta_trans_map), na.rm = T)) %>% 
  select(rsid, chr, pos, a0, a1, beta_EUR_map, beta_AFR_map, 
         beta_trans_map, beta_AD_map)
AD.map.combine.id = AD.map.combine %>% pull(rsid) %>% unique()
length(AD.map.combine.id) # 112
save(AD.map.combine.id, file = paste0(raw_data_path, "modeling/snp_ids/AD.map.combine.id.rda"))
# AD mapped SNPs (restrict to CADD >= 12.37)
AD.map.CADD.combine = beta_merge_raw %>% 
  filter(!is.na(beta_EUR_map_CADD) | !is.na(beta_AFR_map_CADD) | !is.na(beta_trans_map_CADD)) %>% 
  rowwise() %>% 
  mutate(beta_AD_map_CADD = mean(c(beta_EUR_map_CADD, beta_AFR_map_CADD, beta_trans_map_CADD), na.rm = T)) %>% 
  select(rsid, chr, pos, a0, a1, beta_EUR_map_CADD, beta_AFR_map_CADD, 
         beta_trans_map_CADD, beta_AD_map_CADD)
AD.map.CADD.combine.id = AD.map.CADD.combine %>% pull(rsid) %>% unique()
length(AD.map.CADD.combine.id) # 34
save(AD.map.CADD.combine.id, file = paste0(raw_data_path, "modeling/snp_ids/AD.map.CADD.combine.id.rda"))
# Combined ind.sig SNPs
all.indsig.combine = beta_merge_raw %>% 
  filter(!is.na(beta_EUR_indsig) | !is.na(beta_AFR_indsig) | 
           !is.na(beta_trans_indsig) | !is.na(beta_LBD_indsig) |
           !is.na(beta_PD_indsig) | !is.na(beta_PSP_indsig) | 
           !is.na(beta_STROKE_indsig)) %>% left_join(AD.indsig.combine) %>% 
  rowwise() %>% 
  mutate(beta_all_indsig = sum(c(beta_AD_indsig, beta_LBD_indsig, beta_PD_indsig, 
                                 beta_PSP_indsig, beta_STROKE_indsig), na.rm = T))  
all.indsig.combine.id = all.indsig.combine %>% pull(rsid) %>% unique()
length(all.indsig.combine.id) # 190
save(all.indsig.combine.id, file = paste0(raw_data_path, "modeling/snp_ids/all.indsig.combine.id.rda"))
# Combined lead SNPs
all.lead.combine = beta_merge_raw %>% 
  filter(!is.na(beta_EUR_lead) | !is.na(beta_AFR_lead) | 
           !is.na(beta_trans_lead) | !is.na(beta_LBD_lead) |
           !is.na(beta_PD_lead) | !is.na(beta_PSP_lead) | 
           !is.na(beta_STROKE_lead)) %>% left_join(AD.lead.combine) %>% 
  rowwise() %>% 
  mutate(beta_all_lead = sum(c(beta_AD_lead, beta_LBD_lead, beta_PD_lead, 
                               beta_PSP_lead, beta_STROKE_lead), na.rm = T))
all.lead.combine.id = all.lead.combine %>% pull(rsid) %>% unique()
length(all.lead.combine.id) # 76
save(all.lead.combine.id, file = paste0(raw_data_path, "modeling/snp_ids/all.lead.combine.id.rda"))
# Combined map SNPs
all.map.combine = beta_merge_raw %>% 
  filter(!is.na(beta_EUR_map) | !is.na(beta_AFR_map) | 
           !is.na(beta_trans_map) | !is.na(beta_LBD_map) |
           !is.na(beta_PD_map) | !is.na(beta_PSP_map) | 
           !is.na(beta_STROKE_map)) %>% left_join(AD.map.combine) %>% 
  rowwise() %>% 
  mutate(beta_all_map = sum(c(beta_AD_map, beta_LBD_map, beta_PD_map, 
                              beta_PSP_map, beta_STROKE_map), na.rm = T)) 
all.map.combine.id = all.map.combine %>% pull(rsid) %>% unique()
length(all.map.combine.id) # 179
save(all.map.combine.id, file = paste0(raw_data_path, "modeling/snp_ids/all.map.combine.id.rda"))
# Combined map SNPs (restrict to CADD >= 12.37)
all.map.CADD.combine = beta_merge_raw %>% 
  filter(!is.na(beta_EUR_map_CADD) | !is.na(beta_AFR_map_CADD) | 
           !is.na(beta_trans_map_CADD) | !is.na(beta_LBD_map_CADD) |
           !is.na(beta_PD_map_CADD) | !is.na(beta_PSP_map_CADD) | 
           !is.na(beta_STROKE_map_CADD)) %>% left_join(AD.map.CADD.combine) %>% 
  rowwise() %>% 
  mutate(beta_all_map_CADD = sum(c(beta_AD_map_CADD, beta_LBD_map_CADD, beta_PD_map_CADD, 
                              beta_PSP_map_CADD, beta_STROKE_map_CADD), na.rm = T)) 
all.map.CADD.combine.id = all.map.CADD.combine %>% pull(rsid) %>% unique()
length(all.map.CADD.combine.id) # 82
save(all.map.CADD.combine.id, file = paste0(raw_data_path, "modeling/snp_ids/all.map.CADD.combine.id.rda"))
```
