==> 2_sample_selection.Rmd <==
---
title: "2_sample_selection"
author: "Joy_Fu"
date: "2023-05-02"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
rm(list = ls())
pacman::p_load(tidyverse, gtsummary, PheWAS, reshape)
raw_data_path = "/Users/Mingzhou/Desktop/Projects/Dementia-prediction/data/"
output_path = "/Users/Mingzhou/Desktop/Projects/Dementia-prediction/output/"
phe_table_extract_date = "2023-03-20"
# Source in useful functions
source("/Users/Mingzhou/Desktop/Projects/Dementia-prediction/code/functions.R")
```

# Part 1. Raw data preprocessing
## 1. Clean demographic info
```{r message=FALSE, warning=FALSE}
# Read in data
demo_raw = read_csv(file = paste0(raw_data_path, 
                                  "atlas/pheno/raw/atlas_demo_03202023.csv"),
                    col_names = FALSE, na = c("NULL")) 
dim(demo_raw) # dim = (436258,8)
names(demo_raw) = c("PatientID", "UniqueSampleId", "Sex", "Ethnicity", "FirstRace", 
                    "BirthDate", "DeathDate", "PatientLivingStatus")
# Clean demographic info
ddr_demo_short = demo_raw %>% 
  filter(Sex != "X" & Sex != "Unknown") %>% 
  dplyr::select(PatientID, UniqueSampleId, BirthDate, DeathDate) %>% 
  unique() %>% filter(!is.na(UniqueSampleId))
dim(ddr_demo_short) # dim = (65479,4)
```

## 2. Clean encounter info
```{r}
# Read in data
enc_raw = read_csv(file = paste0(raw_data_path, 
                                 "atlas/pheno/raw/atlas_encounter_03202023.csv"),
                   col_names = FALSE, na = c("NULL")) 
dim(enc_raw) # dim = (30399114,5)
names(enc_raw) = c("PatientID", "UniqueSampleId", "Sex", "DiagnosisCode",
                   "StartDate")
enc_demo = enc_raw %>% 
  mutate(StartDate = as.Date(StartDate)) %>% 
  inner_join(ddr_demo_short) %>% 
  mutate(age_diagnosis = as.numeric(StartDate - BirthDate)/365.25) %>% 
  filter(!is.na(age_diagnosis)) %>% 
  arrange(PatientID, StartDate) %>% 
  mutate(valid_record = case_when(
    is.na(DeathDate) & StartDate >= BirthDate ~ 1,
    !is.na(DeathDate) & StartDate >= BirthDate & StartDate <= DeathDate ~ 1
  )) %>% filter(valid_record == 1) %>% select(-valid_record)
dim(enc_demo) # dim = (29628667,8)
```

## 3. Add genetic ancestry and EHR features
```{r}
load(file = paste0(raw_data_path, "atlas/geno/gen_ancestry.rda"))
enc_demo_55 = enc_demo %>% 
  inner_join(gen_ancestry) %>% filter(age_diagnosis >= 55) 
dim(enc_demo_55) # dim = (16398957,9)
save(enc_demo_55, file = paste0(raw_data_path, "atlas/pheno/mod/enc_demo_55.rda"))
# Map to phecodes
patient_ehr55_phecode = enc_demo_55 %>%
  arrange(PatientID, StartDate) %>% group_by(PatientID) %>% 
  mutate(age_last_visit = as.numeric(max(StartDate) - BirthDate)/365.25) %>% 
  mutate(record_length = as.numeric(max(StartDate) - min(StartDate))/365.25) %>% 
  mutate(n_encounter = length(unique(StartDate))) %>% ungroup() %>% 
  mutate(vocabulary_id = "ICD10CM") %>% 
  left_join(PheWAS::phecode_map, by = c("vocabulary_id" = "vocabulary_id",
                                        "DiagnosisCode" = "code")) %>% 
  select(PatientID, UniqueSampleId, Sex, gen_ancestry, age_last_visit,
         record_length, n_encounter, StartDate, phecode) %>% 
  unique() %>% drop_na() %>% group_by(PatientID) %>% 
  # define n_diagnosis by phecodes
  mutate(n_diagnosis = length(unique(phecode))) %>% ungroup() %>% 
  select(PatientID, UniqueSampleId, Sex, gen_ancestry, age_last_visit,
         record_length, n_diagnosis, n_encounter) %>% unique() %>% 
  filter(age_last_visit >= 55 & age_last_visit < 90) %>% filter(n_encounter >= 2) 
dim(patient_ehr55_phecode) # dim = (29666,8)
save(patient_ehr55_phecode, file = paste0(raw_data_path, "atlas/pheno/mod/patient_ehr55_phecode.rda"))
```

## 4. Dementia phenotypes
```{r}
enc_demo_map = enc_demo %>% 
  filter(UniqueSampleId %in% patient_ehr55_phecode$UniqueSampleId) %>% 
  group_by(UniqueSampleId, DiagnosisCode) %>% 
  arrange(StartDate) %>% slice_head(n = 1) %>% ungroup() %>% 
  mutate(vocabulary_id = "ICD10CM") %>% 
  left_join(PheWAS::phecode_map, by = c("vocabulary_id" = "vocabulary_id",
                                        "DiagnosisCode" = "code")) %>% unique()
dim(enc_demo_map) # dim = (3292594,10)
save(enc_demo_map, file = paste0(raw_data_path, "atlas/pheno/mod/enc_demo_map.rda"))
# Dementia cases info
dem_phecode = c("290.1", "290.11", "290.12", "290.13", "290.16")
dem_icds = c("F01", "F01.5", "F01.50", "F01.51", "F01.511", "F01.518",
             "F02.80", "F02.81", "F02.811", "F02.818", 
             "F03", "F03.9", "F03.90", "F03.91", "F03.911", "F03.918",
             "G30", "G30.0", "G30.1", "G30.8", "G30.9",
             "G31.0", "G31.01", "G31.09", "G31.1", "G31.83", "G31.85", 
             "G23.1")
dem_case_info = enc_demo_map %>% 
  # select cases only
  filter(phecode %in% dem_phecode | DiagnosisCode %in% dem_icds) %>% 
  # Collapse to first phecode encounter
  group_by(UniqueSampleId, phecode) %>% 
  arrange(StartDate) %>% slice_head(n = 1) %>% ungroup() %>% 
  select(-DiagnosisCode) %>% unique() %>% group_by(UniqueSampleId) %>% 
  mutate(dem_age_diagnosis = min(age_diagnosis),
         DementiaDate = min(StartDate)) %>% 
  select(PatientID, UniqueSampleId, DementiaDate, dem_age_diagnosis) %>% 
  unique() %>% drop_na()
dim(dem_case_info) # dim = (1200,4)

# Cases
case_ehr = enc_demo_55 %>% inner_join(dem_case_info) %>% 
  mutate(dementia = if_else(!is.na(DementiaDate), 1, 0)) %>% 
  filter(StartDate <= DementiaDate) %>% 
  arrange(PatientID, StartDate) %>% group_by(PatientID) %>% 
  mutate(age_last_visit = as.numeric(max(StartDate) - BirthDate)/365.25) %>% 
  mutate(record_length = as.numeric(max(StartDate) - min(StartDate))/365.25) %>% 
  mutate(n_encounter = length(unique(StartDate))) %>% ungroup() %>% 
  mutate(vocabulary_id = "ICD10CM") %>% 
  left_join(PheWAS::phecode_map, by = c("vocabulary_id" = "vocabulary_id",
                                        "DiagnosisCode" = "code")) %>% 
  select(PatientID, UniqueSampleId, Sex, gen_ancestry, age_last_visit,
         record_length, n_encounter, StartDate, phecode, dementia) %>% ungroup() %>% 
  unique() %>% drop_na() %>% group_by(PatientID) %>% 
  # define n_diagnosis by phecodes
  mutate(n_diagnosis = length(unique(phecode))) %>% ungroup() %>% 
  select(PatientID, UniqueSampleId, Sex, gen_ancestry, age_last_visit,
         record_length, n_diagnosis, n_encounter, dementia) %>% unique() %>% 
  mutate(enc_per_yr = n_encounter/record_length) %>% 
  mutate(female = if_else(Sex == "Female", 1, 0)) %>% filter(n_encounter >= 2) %>% 
  select(UniqueSampleId, gen_ancestry, age_last_visit, female, n_encounter, n_diagnosis,
         record_length, enc_per_yr, dementia) %>% unique() %>% drop_na()
dim(case_ehr) # dim = (1077,9)
# Controls
icd_exclude_tbl = phecode_map_icd10 %>% 
  left_join(phecode_exclude, by = c("phecode" = "code"))
exclude_phecode = icd_exclude_tbl %>% 
  filter(code %in% dem_icds | phecode %in% dem_phecode) %>% 
  pull(exclusion_criteria) %>% unique()
exclude_icd = icd_exclude_tbl %>% filter(phecode %in% exclude_phecode) %>% 
  pull(code) %>% unique()
length(exclude_icd) # 256

exclude_patients = enc_demo_55 %>% 
  filter(UniqueSampleId %!in% case_ehr$UniqueSampleId) %>% 
  filter(DiagnosisCode %in% exclude_icd) %>% pull(UniqueSampleId) %>% unique()
length(exclude_patients) # 6840
control_ehr = patient_ehr55_phecode %>% left_join(dem_case_info) %>% 
  mutate(dementia = if_else(!is.na(DementiaDate), 1, 0)) %>% 
  filter(UniqueSampleId %!in% exclude_patients) %>%
  filter(age_last_visit >= 70 & age_last_visit < 90) %>% 
  filter(n_encounter >= 2) %>% filter(dementia == 0) %>% 
  mutate(enc_per_yr = n_encounter/record_length) %>% 
  mutate(female = if_else(Sex == "Female", 1, 0)) %>% 
  filter(record_length >= 5 & enc_per_yr >= 1) %>%
  select(UniqueSampleId, gen_ancestry, age_last_visit, female, n_encounter, n_diagnosis,
         record_length, enc_per_yr, dementia) %>% unique() %>% drop_na()
dim(control_ehr) # dim = (7103,9)
```

```{r}
final_eligible = rbind(case_ehr, control_ehr) %>% as.data.frame() %>% unique()
dim(final_eligible) # dim = (8180,9)
save(final_eligible, 
     file = paste0(paste0(raw_data_path, "atlas/pheno/mod/final_eligible.rda")))
```


# Part 2. Ancestry samples
```{r}
# load in PCs info
load(file = paste0(raw_data_path, "atlas/geno/pcs/anc_pcs.rda"))
dim(anc_pcs) # dim = (51247, 22)
```

## 1. AMR
```{r}
# final sample
final_df_AMR = final_eligible %>% filter(gen_ancestry == "AMR")
# add PCs info
sample_AMR_full = final_df_AMR %>% 
  dplyr::rename("age" = "age_last_visit") %>% 
  left_join(anc_pcs) %>% select(-c(gen_ancestry))
dim(sample_AMR_full) # dim = (610,28)
save(sample_AMR_full, file = paste0(raw_data_path, "modeling/AMR/sample_AMR_full.rda"))
```

```{r}
# Check distribution in dementia cases/non-cases
sample_AMR_full %>% 
  select(dementia, age, female, record_length, n_encounter, enc_per_yr, n_diagnosis) %>% 
  tbl_summary(by = dementia) %>% add_p()
```


## 2. AFR
```{r}
final_df_AFR = final_eligible %>% filter(gen_ancestry == "AFR") 
# add PCs info
sample_AFR_full = final_df_AFR %>% 
  dplyr::rename("age" = "age_last_visit") %>% 
  left_join(anc_pcs) %>% select(-c(gen_ancestry))
dim(sample_AFR_full) # dim = (440,28)
save(sample_AFR_full, file = paste0(raw_data_path, "modeling/AFR/sample_AFR_full.rda"))
```

```{r}
# Check distribution in dementia cases/non-cases
sample_AFR_full %>% 
  select(dementia, age, female, record_length, n_encounter, enc_per_yr, n_diagnosis) %>% 
  tbl_summary(by = dementia) %>% add_p()
```

## 3. EAS
```{r}
final_df_EAS = final_eligible %>% filter(gen_ancestry == "EAS") 
# add PCs info
load(file = paste0(raw_data_path, "atlas/geno/pcs/eas_pcs.rda"))
sample_EAS_full = final_df_EAS %>% 
  dplyr::rename("age" = "age_last_visit") %>% 
  left_join(anc_pcs) %>% select(-c(gen_ancestry))
dim(sample_EAS_full) # dim = (673,28)
save(sample_EAS_full, file = paste0(raw_data_path, "modeling/EAS/sample_EAS_full.rda"))
```

```{r}
# Check distribution in dementia cases/non-cases
sample_EAS_full %>% 
  select(dementia, age, female, record_length, n_encounter, enc_per_yr, n_diagnosis) %>% 
  tbl_summary(by = dementia) %>% add_p()
```

# Part 3. Donut plot
```{r}
load(file = paste0(raw_data_path, "atlas/pheno/mod/enc_demo_map.rda"))
# Dementia cases info
dem_phecode = c("290.1", "290.11", "290.12", "290.13", "290.16")
dem_icds = c("F01", "F01.5", "F01.50", "F01.51", "F01.511", "F01.518",
             "F02.80", "F02.81", "F02.811", "F02.818", 
             "F03", "F03.9", "F03.90", "F03.91", "F03.911", "F03.918",
             "G30", "G30.0", "G30.1", "G30.8", "G30.9",
             "G31.0", "G31.01", "G31.09", "G31.1", "G31.83", "G31.85", 
             "G23.1")
dem_case_info = enc_demo_map %>% 
  # select cases only
  filter(phecode %in% dem_phecode | DiagnosisCode %in% dem_icds) %>% 
  mutate(ICD_3digit = substr(DiagnosisCode, 1, 3)) %>% 
  select(UniqueSampleId, ICD_3digit) %>% unique()
```

## 1. AMR
```{r}
load(file = paste0(raw_data_path, "modeling/AMR/sample_AMR_full.rda"))
case_AMR = dem_case_info %>% 
  filter(UniqueSampleId %in% sample_AMR_full$UniqueSampleId) %>% 
  mutate(value = 1) %>% 
  spread(key = "ICD_3digit", value = "value") %>% 
  mutate(final_diagnosis = case_when(
    !is.na(G30) ~ "G30",
    !is.na(F01) ~ "F01",
    !is.na(G31) ~ "G31",
    !is.na(F02) ~ "F02",
    !is.na(F03) ~ "F03"
  )) %>% column_to_rownames(var = "UniqueSampleId")

data_plot = table(case_AMR$final_diagnosis) %>% t() %>% as.data.frame() %>% 
  select(-Var1) %>% mutate(Freq = as.numeric(Freq))
names(data_plot) = c("category", "count")
# Compute percentages
data_plot_add = data_plot %>% 
  mutate(fraction = count/sum(count),
         ymax = cumsum(fraction),
         ymin = c(0, head(ymax, n = -1)),
         labelPosition = (ymax + ymin)/2,
         label = paste0(category, "\n count: ", count))
# Make the plot
pdf(paste0(output_path, "donut_AMR.pdf"), width = 6, height = 5)
ggplot(data_plot_add, 
       aes(ymax = ymax, ymin = ymin, xmax = 4, xmin = 3, fill = category)) +
  geom_rect() + 
  geom_text(x = 2, aes(y = labelPosition, label = label, color = category), size = 4) +
  scale_fill_brewer(palette = "Set2") +
  scale_color_brewer(palette = "Set2") +
  coord_polar(theta = "y") +
  xlim(c(-1, 4)) +
  theme_void() +
  theme(legend.position = "none")
dev.off()
```

## 2. AFR
```{r}
load(file = paste0(raw_data_path, "modeling/AFR/sample_AFR_full.rda"))
case_AFR = dem_case_info %>% 
  filter(UniqueSampleId %in% sample_AFR_full$UniqueSampleId) %>% 
  mutate(value = 1) %>% 
  spread(key = "ICD_3digit", value = "value") %>% 
  mutate(final_diagnosis = case_when(
    !is.na(G30) ~ "G30",
    !is.na(F01) ~ "F01",
    !is.na(G31) ~ "G31",
    !is.na(F02) ~ "F02",
    !is.na(F03) ~ "F03"
  )) %>% column_to_rownames(var = "UniqueSampleId")

data_plot = table(case_AFR$final_diagnosis) %>% t() %>% as.data.frame() %>% 
  select(-Var1) %>% mutate(Freq = as.numeric(Freq))
names(data_plot) = c("category", "count")
# Compute percentages
data_plot_add = data_plot %>% 
  mutate(fraction = count/sum(count),
         ymax = cumsum(fraction),
         ymin = c(0, head(ymax, n = -1)),
         labelPosition = (ymax + ymin)/2,
         label = paste0(category, "\n count: ", count))
# Make the plot
pdf(paste0(output_path, "donut_AFR.pdf"), width = 6, height = 5)
ggplot(data_plot_add, 
       aes(ymax = ymax, ymin = ymin, xmax = 4, xmin = 3, fill = category)) +
  geom_rect() + 
  geom_text(x = 2, aes(y = labelPosition, label = label, color = category), size = 4) +
  scale_fill_brewer(palette = "Set2") +
  scale_color_brewer(palette = "Set2") +
  coord_polar(theta = "y") +
  xlim(c(-1, 4)) +
  theme_void() +
  theme(legend.position = "none")
dev.off()
```

## 3. EAS
```{r}
load(file = paste0(raw_data_path, "modeling/EAS/sample_EAS_full.rda"))
case_EAS = dem_case_info %>% 
  filter(UniqueSampleId %in% sample_EAS_full$UniqueSampleId) %>% 
  mutate(value = 1) %>% 
  spread(key = "ICD_3digit", value = "value") %>% 
  mutate(final_diagnosis = case_when(
    !is.na(G30) ~ "G30",
    !is.na(F01) ~ "F01",
    !is.na(G31) ~ "G31",
    !is.na(F02) ~ "F02",
    !is.na(F03) ~ "F03"
  )) %>% column_to_rownames(var = "UniqueSampleId")

data_plot = table(case_EAS$final_diagnosis) %>% t() %>% as.data.frame() %>% 
  select(-Var1) %>% mutate(Freq = as.numeric(Freq))
names(data_plot) = c("category", "count")
# Compute percentages
data_plot_add = data_plot %>% 
  mutate(fraction = count/sum(count),
         ymax = cumsum(fraction),
         ymin = c(0, head(ymax, n = -1)),
         labelPosition = (ymax + ymin)/2,
         label = paste0(category, "\n count: ", count))
# Make the plot
pdf(paste0(output_path, "donut_EAS.pdf"), width = 6, height = 5)
ggplot(data_plot_add, 
       aes(ymax = ymax, ymin = ymin, xmax = 4, xmin = 3, fill = category)) +
  geom_rect() + 
  geom_text(x = 2, aes(y = labelPosition, label = label, color = category), size = 4) +
  scale_fill_brewer(palette = "Set2") +
  scale_color_brewer(palette = "Set2") +
  coord_polar(theta = "y") +
  xlim(c(-1, 4)) +
  theme_void() +
  theme(legend.position = "none")
dev.off()
```

# Part 4. Pairwise PRS associations
```{r}
load(file = paste0(raw_data_path, "prs/mod/atlas_prs_final.rda"))
load(file = paste0(raw_data_path, "modeling/AMR/sample_AMR_full.rda"))
load(file = paste0(raw_data_path, "modeling/AFR/sample_AFR_full.rda"))
load(file = paste0(raw_data_path, "modeling/EAS/sample_EAS_full.rda"))
```

## 1. AMR
```{r}
sample_AMR_prs = sample_AMR_full %>% 
  mutate(UniqueSampleId = as.character(UniqueSampleId)) %>% 
  left_join(atlas_prs_final) %>% column_to_rownames(var = "UniqueSampleId") %>% 
  select(-c(gen_ancestry, APOE, e2count)) %>% drop_na()
dim(sample_AMR_prs) # dim = (610,56)
# select numeric variables
amr_corr = sample_AMR_prs %>% select(contains("indsig"), contains("map"))
cor_amr = round(cor(amr_corr, use = "complete.obs"), 2)
upper_tri = get_upper_tri(cor_amr)
melted_cormat = melt(cor_amr, na.rm = T)
write.table(melted_cormat, 
            file = paste0(output_path, "AMR_PRS_corr.txt"),
            sep = "\t", quote = F, row.names = F, col.names = T)
```

## 2. AFR
```{r}
sample_AFR_prs = sample_AFR_full %>% 
  mutate(UniqueSampleId = as.character(UniqueSampleId)) %>% 
  left_join(atlas_prs_final) %>% column_to_rownames(var = "UniqueSampleId") %>% 
  select(-c(gen_ancestry, APOE, e2count)) %>% drop_na()
dim(sample_AFR_prs) # dim = (440,56)
# select numeric variables
afr_corr = sample_AFR_prs %>% select(contains("indsig"), contains("map"))
cor_afr = round(cor(afr_corr, use = "complete.obs"), 2)
upper_tri = get_upper_tri(cor_afr)
melted_cormat = melt(cor_afr, na.rm = T)
write.table(melted_cormat, 
            file = paste0(output_path, "AFR_PRS_corr.txt"),
            sep = "\t", quote = F, row.names = F, col.names = T)
```
