==> 3_modeling.Rmd <==
---
title: "3_modeling"
author: "Joy_Fu"
date: "2023-05-02"
output: html_document
---

```{r message=FALSE, warning=FALSE}
rm(list = ls())
# lapply(paste('package:', names(sessionInfo()$otherPkgs), sep = ""),
#        detach, character.only = TRUE, unload = TRUE)
pacman::p_load(tidyverse, h2o, caret, pROC, forestplot)
code_path = "/Users/Mingzhou/Desktop/Projects/Dementia-prediction/code/"
raw_data_path = "/Users/Mingzhou/Desktop/Projects/Dementia-prediction/data/"
output_path = "/Users/Mingzhou/Desktop/Projects/Dementia-prediction/output/"
# Source in useful functions
source("/Users/Mingzhou/Desktop/Projects/Dementia-prediction/code/functions.R")
# Read in SNP.id data
ids_folder = paste0(raw_data_path, "modeling/snp_ids") 
ids_files = list.files(ids_folder, pattern = ".rda$")
for (i in ids_files) {
  load(paste0(raw_data_path, 'modeling/snp_ids/', i))
}
```

# Part 1. AMR analyses
We ran bootstrapping (x1000) to get PRS model results.
```{r}
# source(paste0(code_path, "boot_PRS.R"))
# load in results
load(file = paste0(raw_data_path, "modeling/AMR/PRS_bootstrap_full.rda"))
dim(PRS_bootstrap_full) # dim = (16000,30)
na_counts = PRS_bootstrap_full %>% unique() %>%
  summarise(across(everything(), ~sum(is.na(.)))) # no NAs
```

We ran bootstrapping (x1000) to get SNP model results.
```{r}
# h2o.shutdown(prompt = F)
# h2o.init(nthreads = -1)
# source(paste0(code_path, "boot_SNP.R"))
# load in results
load(file = paste0(raw_data_path, "modeling/AMR/SNP_bootstrap_full.rda"))
dim(SNP_bootstrap_full) # dim = (8000,30)
na_counts = SNP_bootstrap_full %>% unique() %>%
  summarise(across(everything(), ~sum(is.na(.)))) # no NAs
```

## 1. Summary table
```{r}
# Specify the desired order of the character column
desired_order = c("e4count", "PRS.EUR.indsig", "PRS.EUR.map", "PRS.AFR.indsig", 
                  "PRS.AFR.map", "PRS.Trans.indsig", "PRS.Trans.map",
                  "PRS.AD.indsig.combine", "PRS.AD.map.combine", 
                  "PRS.neuro.indsig.combine", "PRS.neuro.map.combine",
                  "SNP.AD.indsig", "SNP.AD.map", "SNP.all.indsig",
                  "SNP.all.map")
# desired_order = c("PRS.EUR.lead", "PRS.AFR.lead", "PRS.Trans.lead",
#                   "PRS.AD.lead.combine", "PRS.neuro.lead.combine",
#                   "SNP.AD.lead", "SNP.all.lead")
```

```{r}
prs_summary_tbl = PRS_bootstrap_full %>% 
  select(AUCPR, AUC, f1, accuracy, precision, recall, specificity, 
         mark, bootstrap) %>% group_by(mark) %>% 
  summarise(mean_aucpr = mean(AUCPR), lower_aucpr = quantile(AUCPR, 0.025), 
            upper_aucpr = quantile(AUCPR, 0.975), mean_auc = mean(AUC), 
            lower_auc = quantile(AUC, 0.025), upper_auc = quantile(AUC, 0.975),
            mean_f1 = mean(f1), lower_f1 = quantile(f1, 0.025), 
            upper_f1 = quantile(f1, 0.975), mean_accuracy = mean(accuracy), 
            lower_accuracy = quantile(accuracy, 0.025), 
            upper_accuracy = quantile(accuracy, 0.975),
            mean_precision = mean(precision),
            lower_precision = quantile(precision, 0.025), 
            upper_precision = quantile(precision, 0.975),
            mean_recall = mean(recall), 
            lower_recall = quantile(recall, 0.025),
            upper_recall = quantile(recall, 0.975),
            mean_specificity = mean(specificity),
            lower_specificity = quantile(specificity, 0.025),
            upper_specificity = quantile(specificity, 0.975)) %>%
  mutate(AUCPR_stats = paste0(round(mean_aucpr, 3), " (", 
                 round(lower_aucpr, 3), ", ", round(upper_aucpr, 3), ")"),
         AUC_stats = paste0(round(mean_auc, 3), " (", 
                 round(lower_auc, 3), ", ", round(upper_auc, 3), ")"),
         f1_stats = paste0(round(mean_f1, 3), " (", 
                 round(lower_f1, 3), ", ", round(upper_f1, 3), ")"),
         accuracy_stats = paste0(round(mean_accuracy, 3), " (", 
                 round(lower_accuracy, 3), ", ", round(upper_accuracy, 3), ")"),
         precision_stats = paste0(round(mean_precision, 3), " (",
                 round(lower_precision, 3), ", ", 
                 round(upper_precision, 3), ")"),
         recall_stats = paste0(round(mean_recall, 3), " (",
                 round(lower_recall, 3), ", ", round(upper_recall, 3), ")"),
         specificity_stats = paste0(round(mean_specificity, 3), " (",
                 round(lower_specificity, 3), ", ", 
                 round(upper_specificity, 3), ")")) %>%
  select(mark, AUCPR_stats, AUC_stats, f1_stats, accuracy_stats, precision_stats, 
         recall_stats, specificity_stats) 

snp_summary_tbl = SNP_bootstrap_full %>% 
  select(AUCPR, AUC, f1, accuracy, precision, recall, specificity, 
         mark, bootstrap) %>% group_by(mark) %>% 
  summarise(mean_AUCPR = mean(AUCPR), lower_AUCPR = quantile(AUCPR, 0.025), 
            upper_AUCPR = quantile(AUCPR, 0.975), mean_AUC = mean(AUC), 
            lower_AUC = quantile(AUC, 0.025), upper_AUC = quantile(AUC, 0.975),
            mean_f1 = mean(f1), lower_f1 = quantile(f1, 0.025), 
            upper_f1 = quantile(f1, 0.975), mean_accuracy = mean(accuracy), 
            lower_accuracy = quantile(accuracy, 0.025), 
            upper_accuracy = quantile(accuracy, 0.975),
            mean_precision = mean(precision),
            lower_precision = quantile(precision, 0.025), 
            upper_precision = quantile(precision, 0.975),
            mean_recall = mean(recall), 
            lower_recall = quantile(recall, 0.025),
            upper_recall = quantile(recall, 0.975),
            mean_specificity = mean(specificity),
            lower_specificity = quantile(specificity, 0.025),
            upper_specificity = quantile(specificity, 0.975)) %>%
  mutate(AUCPR_stats = paste0(round(mean_AUCPR, 3), " (", 
                 round(lower_AUCPR, 3), ", ", round(upper_AUCPR, 3), ")"),
         AUC_stats = paste0(round(mean_AUC, 3), " (", 
                 round(lower_AUC, 3), ", ", round(upper_AUC, 3), ")"),
         f1_stats = paste0(round(mean_f1, 3), " (", 
                 round(lower_f1, 3), ", ", round(upper_f1, 3), ")"),
         accuracy_stats = paste0(round(mean_accuracy, 3), " (", 
                 round(lower_accuracy, 3), ", ", round(upper_accuracy, 3), ")"),
         precision_stats = paste0(round(mean_precision, 3), " (",
                 round(lower_precision, 3), ", ", 
                 round(upper_precision, 3), ")"),
         recall_stats = paste0(round(mean_recall, 3), " (",
                 round(lower_recall, 3), ", ", round(upper_recall, 3), ")"),
         specificity_stats = paste0(round(mean_specificity, 3), " (",
                 round(lower_specificity, 3), ", ", 
                 round(upper_specificity, 3), ")")) %>%
  select(mark, AUCPR_stats, AUC_stats, f1_stats, accuracy_stats, precision_stats, 
         recall_stats, specificity_stats)
```

```{r}
AMR_summary_tbl = rbind(prs_summary_tbl, snp_summary_tbl) %>% 
  filter(mark %in% desired_order) %>% 
  arrange(factor(mark, levels = desired_order))
# output to csv
write.csv(AMR_summary_tbl, 
          paste0(output_path, "model_stats/AMR_summary_tbl.csv") ,
          row.names = FALSE)
# write.csv(AMR_summary_tbl,
#           paste0(output_path, "model_stats/AMR_summary_tbl_lead.csv") ,
#           row.names = FALSE)
```


## 2. Forest plot
### 1) AUCPR
```{r}
prs_summary_tbl = PRS_bootstrap_full %>% 
  select(AUCPR, mark, bootstrap) %>% group_by(mark) %>% 
  summarise(mean_aucpr = mean(AUCPR), 
            lower_CI = quantile(AUCPR, 0.025),
            upper_CI = quantile(AUCPR, 0.975)) 
snp_summary_tbl = SNP_bootstrap_full %>% 
  select(AUCPR, mark, bootstrap) %>% group_by(mark) %>% 
  summarise(mean_aucpr = mean(AUCPR), 
            lower_CI = quantile(AUCPR, 0.025),
            upper_CI = quantile(AUCPR, 0.975)) 
AMR_summary_tbl = rbind(prs_summary_tbl, snp_summary_tbl) %>% 
  filter(mark %in% desired_order) %>% 
  arrange(factor(mark, levels = desired_order))
# Add CI to the table
ci_text = paste0(round(AMR_summary_tbl$mean_aucpr, 3), " (", 
                 round(AMR_summary_tbl$lower_CI, 3), ", ", 
                 round(AMR_summary_tbl$upper_CI, 3), ")")
```

```{r}
# build a forest plot 
forest.data = AMR_summary_tbl %>% 
  dplyr::select(mean_aucpr, lower_CI, upper_CI) %>% 
  mutate(sub = c("e4count",
    'EUR single PRS', rep(NA, 1), 
    'AFR single PRS' , rep(NA, 1),
    'Transethnic single PRS', rep(NA, 1),
    'AD combined PRSs', rep(NA, 1),
    'Neuro combined PRSs', rep(NA, 1),
    'AD SNPs', rep(NA, 1),
    'Neuro PRSs', rep(NA, 1))) 
tabletext = cbind(
  c("Category", "\n", forest.data$sub),
  c("Predictor", "\n", AMR_summary_tbl$mark),
  c("AUCPR (95% CI)", "\n", ci_text)
)
pop.cols = c("black","black","black")
```

```{r}
pdf(paste0(output_path, "AMR_AUCPR_Forest.pdf"), width = 8, height = 4.5)
forestplot(labeltext = tabletext, graph.pos = 1,
           mean = c(NA, NA, forest.data$mean_aucpr), 
           lower = c(NA, NA, forest.data$lower_CI),
           upper = c(NA, NA, forest.data$upper_CI),
           
           xticks = c(0.20, 0.30, 0.40, 0.50, 0.60),
           zero = max(forest.data$mean_aucpr),
           title = "Figure 2. Forest plot of models' AUCPR for AMR",
           xlab = "AUPRC with 95% CIs from bootstrapping",
           
           txt_gp = fpTxtGp(label = list(gpar(fontface = "bold", cex = 0.8, 
                                              fontfamily = "serif"),
                                         gpar(cex = 0.8, fontfamily = "serif"),
                                         gpar(cex = 0.8, fontfamily = "serif")),
                            ticks = gpar(cex = 0.6, fontfamily = "serif"),
                            xlab = gpar(cex = 0.7, fontfamily = "serif"),
                            title = gpar(cex = 1, fontfamily = "serif")),
           col = fpColors(text = pop.cols[c(1, 1, forest.data$class)], 
                          box  ="black",
                          lines = "black", 
                          zero ="gray50"),
           
           cex = 0.2, lineheight = "auto", boxsize = 0.25, 
           lwd.ci = 1, ci.vertices = TRUE, ci.vertices.height = 0.15)
dev.off()
```

### 2) AUC
```{r}
prs_summary_tbl = PRS_bootstrap_full %>% 
  select(AUC, mark, bootstrap) %>% group_by(mark) %>% 
  summarise(mean_auc = mean(AUC), 
            lower_CI = quantile(AUC, 0.025),
            upper_CI = quantile(AUC, 0.975)) 
snp_summary_tbl = SNP_bootstrap_full %>% 
  select(AUC, mark, bootstrap) %>% group_by(mark) %>% 
  summarise(mean_auc = mean(AUC), 
            lower_CI = quantile(AUC, 0.025),
            upper_CI = quantile(AUC, 0.975)) 
AMR_summary_tbl = rbind(prs_summary_tbl, snp_summary_tbl) %>% 
  filter(mark %in% desired_order) %>% 
  arrange(factor(mark, levels = desired_order))
# Add CI to the table
ci_text = paste0(round(AMR_summary_tbl$mean_auc, 3), " (", 
                 round(AMR_summary_tbl$lower_CI, 3), ", ", 
                 round(AMR_summary_tbl$upper_CI, 3), ")")
```

```{r}
# build a forest plot 
forest.data = AMR_summary_tbl %>% 
  dplyr::select(mean_auc, lower_CI, upper_CI) %>% 
  mutate(sub = c("e4count",
    'EUR single PRS', rep(NA, 1), 
    'AFR single PRS' , rep(NA, 1),
    'Transethnic single PRS', rep(NA, 1),
    'AD combined PRSs', rep(NA, 1),
    'Neuro combined PRSs', rep(NA, 1),
    'AD SNPs', rep(NA, 1),
    'Neuro PRSs', rep(NA, 1))) 
tabletext = cbind(
  c("Category", "\n", forest.data$sub),
  c("Predictor", "\n", AMR_summary_tbl$mark),
  c("AUC (95% CI)", "\n", ci_text)
)
pop.cols = c("black","black","black")
```

```{r}
pdf(paste0(output_path, "AMR_AUC_Forest.pdf"), width = 8, height = 4.5)
forestplot(labeltext = tabletext, graph.pos = 1,
           mean = c(NA, NA, forest.data$mean_auc), 
           lower = c(NA, NA, forest.data$lower_CI),
           upper = c(NA, NA, forest.data$upper_CI),
           
           xticks = c(0.4, 0.5, 0.6, 0.7, 0.8),
           zero = max(forest.data$mean_auc),
           title = "Figure 2. Forest plot of models' AUC for AMR",
           xlab = "AUC with 95% CIs from bootstrapping",
           
           txt_gp = fpTxtGp(label = list(gpar(fontface = "bold", cex = 0.8, 
                                              fontfamily = "serif"),
                                         gpar(cex = 0.8, fontfamily = "serif"),
                                         gpar(cex = 0.8, fontfamily = "serif")),
                            ticks = gpar(cex = 0.6, fontfamily = "serif"),
                            xlab = gpar(cex = 0.7, fontfamily = "serif"),
                            title = gpar(cex = 1, fontfamily = "serif")),
           col = fpColors(text = pop.cols[c(1, 1, forest.data$class)], 
                          box  ="black",
                          lines = "black", 
                          zero ="gray50"),
           
           cex = 0.2, lineheight = "auto", boxsize = 0.25, 
           lwd.ci = 1, ci.vertices = TRUE, ci.vertices.height = 0.15)
dev.off()
```

## 3. P-value calculation
```{r}
combine_summary = rbind(PRS_bootstrap_full, SNP_bootstrap_full) 
dim(combine_summary) # dim = (24000, 30)

prs_columns = c("PRS.EUR.indsig", "PRS.EUR.map", "PRS.AFR.indsig", 
                "PRS.AFR.map", "PRS.Trans.indsig", "PRS.Trans.map",
                "PRS.AD.indsig.combine", "PRS.AD.map.combine",
                "PRS.neuro.indsig.combine", "PRS.neuro.map.combine")
snp_columns = c("SNP.AD.indsig", "SNP.AD.map", "SNP.all.indsig",
                "SNP.all.map")
aucpr_p = combine_summary %>%  
  select(mark, bootstrap, AUCPR) %>%
  pivot_wider(names_from = mark, values_from = AUCPR)
aucpr_p$max_prs = pmap_dbl(aucpr_p[prs_columns], max)
aucpr_p$max_snp = pmap_dbl(aucpr_p[snp_columns], max)
aucpr_p_summary = aucpr_p %>% 
  select(bootstrap, e4count, max_prs, max_snp, SNP.all.map) %>%
  mutate(prs_p = if_else(max_prs < SNP.all.map, 1, 0),
         e4count_p = if_else(e4count < SNP.all.map, 1, 0),
         map_is_max = if_else(max_snp <= SNP.all.map & max_prs < SNP.all.map, 1, 0)) 

table(aucpr_p_summary$e4count_p)[1]/1000                          
table(aucpr_p_summary$prs_p)[1]/1000
table(aucpr_p_summary$map_is_max)[1]/1000
```

```{r}
auc_p = combine_summary %>%  
  select(mark, bootstrap, AUC) %>%
  pivot_wider(names_from = mark, values_from = AUC)
auc_p$max_prs = pmap_dbl(auc_p[prs_columns], max)
auc_p$max_snp = pmap_dbl(auc_p[snp_columns], max)
auc_p_summary = auc_p %>% 
  select(bootstrap, e4count, max_prs, max_snp, SNP.all.map) %>%
  mutate(prs_p = if_else(max_prs < SNP.all.map, 1, 0),
         e4count_p = if_else(e4count < SNP.all.map, 1, 0),
         map_is_max = if_else(max_snp <= SNP.all.map & max_prs < SNP.all.map, 1, 0)) 

table(auc_p_summary$e4count_p)[1]/1000                          
table(auc_p_summary$prs_p)[1]/1000
table(auc_p_summary$map_is_max)[1]/1000
```


## 4. Selected SNPs
We ran bootstrapping (x1000) to get SNP map model (feature selected with 95%
confidence).
```{r}
# source(paste0(code_path, "snp_all_map_selection.R"))
load(file = paste0(raw_data_path, "modeling/AMR/feature_map_full.rda"))
max_boot = max(feature_map_full$bootstrap)
snps_map_summary = feature_map_full %>% 
  group_by(variable) %>% 
  summarise(n = n(), mean_imp = mean(percentage, na.rm = T), 
            lower_imp = quantile(percentage, 0.025, na.rm = T),
            upper_imp = quantile(percentage, 0.975, na.rm = T)) %>%
  arrange(desc(mean_imp)) %>% filter(n >= max_boot * 0.95) %>% 
  mutate(var_imp = paste0(round(mean_imp, 3), " (", 
                 round(lower_imp, 3), ", ", round(upper_imp, 3), ")")) %>% 
  select(variable, var_imp) %>% filter(variable != "Intercept")
dim(snps_map_summary) # dim = (15,2)
```

```{r}
load(file = paste0(raw_data_path, "FUMA_outputs/overlap_fuma.rda"))
load(file = paste0(raw_data_path, "Interpretation/info_tbl/snp_info_full.rda"))
AMR_SNP_map = overlap_fuma %>% 
  filter(chr_pos_name %in% snps_map_summary$variable) %>% 
  select(chr_pos_name, rsID, chr_num, pos, phenotype) %>% 
  left_join(snp_info_full, by = c("rsID" = "rsID")) %>% unique() %>% 
  left_join(snps_map_summary, by = c("chr_pos_name" = "variable")) %>% 
  arrange(factor(chr_pos_name, levels = snps_map_summary$variable)) %>% 
  select(var_imp, rsID, chr_num, pos, CADD, nearestGene, dist, func, posMapFilt, 
         eqtlMapFilt, ciMapFilt, phenotype) %>% mutate(mark = "x") %>% 
  pivot_wider(names_from = phenotype, values_from = mark) 
dim(AMR_SNP_map) # dim = (15,18)
write.table(AMR_SNP_map, 
            file = paste0(output_path, "interpretation/AMR_SNP_map.txt"),
            sep = "\t", quote = F, row.names = F, col.names = T)
```


# Part 2. AFR analyses
We ran bootstrapping (x1000) to get PRS model results.
```{r}
# load(file = paste0(raw_data_path, "modeling/AFR/PRS_bootstrap_full.rda"))
# dim(PRS_bootstrap_full) # dim = (21000,30)
# na_counts = PRS_bootstrap_full %>% unique() %>%
#   summarise(across(everything(), ~sum(is.na(.))))
# exclude_boot_id = PRS_bootstrap_full %>% 
#   filter(is.na(f1)) %>% pull(bootstrap) %>% unique()
# length(exclude_boot_id) # 490
# PRS_bootstrap_filter = PRS_bootstrap_full %>% 
#   filter(bootstrap %!in% exclude_boot_id)
# dim(PRS_bootstrap_filter) # dim = (10710,30)
# dim(PRS_bootstrap_filter)[1]/21
# save(PRS_bootstrap_filter, 
#      file = paste0(raw_data_path, "modeling/AFR/PRS_bootstrap_full_mod.rda"))
```

```{r}
load(file = paste0(raw_data_path, "modeling/AFR/PRS_bootstrap_full_mod.rda"))
dim(PRS_bootstrap_filter) # dim = (21000,30)
na_counts = PRS_bootstrap_filter %>% unique() %>%
  summarise(across(everything(), ~sum(is.na(.)))) # no NAs
```

We ran bootstrapping (x1000) to get SNP model results.
```{r}
# h2o.shutdown(prompt = F)
# h2o.init(nthreads = -1)
# Source in code (run background)
# source(paste0(code_path, "boot_SNP.R"))
# load in results
# SNP_bootstrap_7 = SNP_bootstrap_full
# SNP_bootstrap_final = rbind(SNP_bootstrap_1, SNP_bootstrap_2, SNP_bootstrap_3, 
#                             SNP_bootstrap_4, SNP_bootstrap_5, SNP_bootstrap_6, 
#                             SNP_bootstrap_7)
# dim(SNP_bootstrap_final) # dim = (8000,30)
# SNP_bootstrap_full = SNP_bootstrap_final
# save(SNP_bootstrap_full, 
#      file = paste0(raw_data_path, "modeling/AFR/SNP_bootstrap_full.rda"))
load(file = paste0(raw_data_path, "modeling/AFR/SNP_bootstrap_full.rda"))
dim(SNP_bootstrap_full) # dim = (8000,30)
na_counts = SNP_bootstrap_full %>% unique() %>%
  summarise(across(everything(), ~sum(is.na(.))))
```

## 1. Summary table
```{r}
# Specify the desired order of the character column
desired_order = c("e4count", "PRS.EUR.indsig", "PRS.EUR.map", "PRS.AFR.indsig", 
                  "PRS.AFR.map", "PRS.trans.indsig", "PRS.trans.map",
                  "PRS.AD.indsig.combine", "PRS.AD.map.combine", 
                  "PRS.neuro.indsig.combine", "PRS.neuro.map.combine",
                  "SNP.AD.indsig", "SNP.AD.map", "SNP.all.indsig",
                  "SNP.all.map")
# desired_order = c("PRS.EUR.lead", "PRS.AFR.lead", "PRS.trans.lead",
#                   "PRS.AD.lead.combine", "PRS.neuro.lead.combine",
#                   "SNP.AD.lead", "SNP.all.lead")
```

```{r}
prs_summary_tbl = PRS_bootstrap_filter %>% 
  select(AUCPR, AUC, f1, accuracy, precision, recall, specificity, 
         mark, bootstrap) %>% group_by(mark) %>% 
  summarise(mean_aucpr = mean(AUCPR), lower_aucpr = quantile(AUCPR, 0.025), 
            upper_aucpr = quantile(AUCPR, 0.975), mean_auc = mean(AUC), 
            lower_auc = quantile(AUC, 0.025), upper_auc = quantile(AUC, 0.975),
            mean_f1 = mean(f1), lower_f1 = quantile(f1, 0.025), 
            upper_f1 = quantile(f1, 0.975), mean_accuracy = mean(accuracy), 
            lower_accuracy = quantile(accuracy, 0.025), 
            upper_accuracy = quantile(accuracy, 0.975),
            mean_precision = mean(precision),
            lower_precision = quantile(precision, 0.025), 
            upper_precision = quantile(precision, 0.975),
            mean_recall = mean(recall), 
            lower_recall = quantile(recall, 0.025),
            upper_recall = quantile(recall, 0.975),
            mean_specificity = mean(specificity),
            lower_specificity = quantile(specificity, 0.025),
            upper_specificity = quantile(specificity, 0.975)) %>%
  mutate(AUCPR_stats = paste0(round(mean_aucpr, 3), " (", 
                 round(lower_aucpr, 3), ", ", round(upper_aucpr, 3), ")"),
         AUC_stats = paste0(round(mean_auc, 3), " (", 
                 round(lower_auc, 3), ", ", round(upper_auc, 3), ")"),
         f1_stats = paste0(round(mean_f1, 3), " (", 
                 round(lower_f1, 3), ", ", round(upper_f1, 3), ")"),
         accuracy_stats = paste0(round(mean_accuracy, 3), " (", 
                 round(lower_accuracy, 3), ", ", round(upper_accuracy, 3), ")"),
         precision_stats = paste0(round(mean_precision, 3), " (",
                 round(lower_precision, 3), ", ", 
                 round(upper_precision, 3), ")"),
         recall_stats = paste0(round(mean_recall, 3), " (",
                 round(lower_recall, 3), ", ", round(upper_recall, 3), ")"),
         specificity_stats = paste0(round(mean_specificity, 3), " (",
                 round(lower_specificity, 3), ", ", 
                 round(upper_specificity, 3), ")")) %>%
  select(mark, AUCPR_stats, AUC_stats, f1_stats, accuracy_stats, precision_stats, 
         recall_stats, specificity_stats) 

snp_summary_tbl = SNP_bootstrap_full %>% 
  select(AUCPR, AUC, f1, accuracy, precision, recall, specificity, 
         mark, bootstrap) %>% group_by(mark) %>% 
  summarise(mean_AUCPR = mean(AUCPR), lower_AUCPR = quantile(AUCPR, 0.025), 
            upper_AUCPR = quantile(AUCPR, 0.975), mean_AUC = mean(AUC), 
            lower_AUC = quantile(AUC, 0.025), upper_AUC = quantile(AUC, 0.975),
            mean_f1 = mean(f1), lower_f1 = quantile(f1, 0.025), 
            upper_f1 = quantile(f1, 0.975), mean_accuracy = mean(accuracy), 
            lower_accuracy = quantile(accuracy, 0.025), 
            upper_accuracy = quantile(accuracy, 0.975),
            mean_precision = mean(precision),
            lower_precision = quantile(precision, 0.025), 
            upper_precision = quantile(precision, 0.975),
            mean_recall = mean(recall), 
            lower_recall = quantile(recall, 0.025),
            upper_recall = quantile(recall, 0.975),
            mean_specificity = mean(specificity),
            lower_specificity = quantile(specificity, 0.025),
            upper_specificity = quantile(specificity, 0.975)) %>%
  mutate(AUCPR_stats = paste0(round(mean_AUCPR, 3), " (", 
                 round(lower_AUCPR, 3), ", ", round(upper_AUCPR, 3), ")"),
         AUC_stats = paste0(round(mean_AUC, 3), " (", 
                 round(lower_AUC, 3), ", ", round(upper_AUC, 3), ")"),
         f1_stats = paste0(round(mean_f1, 3), " (", 
                 round(lower_f1, 3), ", ", round(upper_f1, 3), ")"),
         accuracy_stats = paste0(round(mean_accuracy, 3), " (", 
                 round(lower_accuracy, 3), ", ", round(upper_accuracy, 3), ")"),
         precision_stats = paste0(round(mean_precision, 3), " (",
                 round(lower_precision, 3), ", ", 
                 round(upper_precision, 3), ")"),
         recall_stats = paste0(round(mean_recall, 3), " (",
                 round(lower_recall, 3), ", ", round(upper_recall, 3), ")"),
         specificity_stats = paste0(round(mean_specificity, 3), " (",
                 round(lower_specificity, 3), ", ", 
                 round(upper_specificity, 3), ")")) %>%
  select(mark, AUCPR_stats, AUC_stats, f1_stats, accuracy_stats, precision_stats, 
         recall_stats, specificity_stats)
```

```{r}
AFR_summary_tbl = rbind(prs_summary_tbl, snp_summary_tbl) %>% 
  filter(mark %in% desired_order) %>% 
  arrange(factor(mark, levels = desired_order))
# output to csv
write.csv(AFR_summary_tbl, 
          paste0(output_path, "model_stats/AFR_summary_tbl.csv") ,
          row.names = FALSE)
# write.csv(AFR_summary_tbl,
#           paste0(output_path, "model_stats/AFR_summary_tbl_lead.csv") ,
#           row.names = FALSE)
```

## 2. Forest plot
### 1) AUCPR
```{r}
prs_summary_tbl = PRS_bootstrap_filter %>% 
  select(AUCPR, mark, bootstrap) %>% group_by(mark) %>% 
  summarise(mean_aucpr = mean(AUCPR), 
            lower_CI = quantile(AUCPR, 0.025),
            upper_CI = quantile(AUCPR, 0.975)) 
snp_summary_tbl = SNP_bootstrap_full %>% 
  select(AUCPR, mark, bootstrap) %>% group_by(mark) %>% 
  summarise(mean_aucpr = mean(AUCPR), 
            lower_CI = quantile(AUCPR, 0.025),
            upper_CI = quantile(AUCPR, 0.975)) 
AFR_summary_tbl = rbind(prs_summary_tbl, snp_summary_tbl) %>% 
  filter(mark %in% desired_order) %>% 
  arrange(factor(mark, levels = desired_order))
# Add CI to the table
ci_text = paste0(round(AFR_summary_tbl$mean_aucpr, 3), " (", 
                 round(AFR_summary_tbl$lower_CI, 3), ", ", 
                 round(AFR_summary_tbl$upper_CI, 3), ")")
```

```{r}
# build a forest plot 
forest.data = AFR_summary_tbl %>% 
  dplyr::select(mean_aucpr, lower_CI, upper_CI) %>% 
  mutate(sub = c("e4count",
    'EUR single PRS', rep(NA, 1), 
    'AFR single PRS' , rep(NA, 1),
    'Transethnic single PRS', rep(NA, 1),
    'AD combined PRSs', rep(NA, 1),
    'Neuro combined PRSs', rep(NA, 1),
    'AD SNPs', rep(NA, 1),
    'Neuro PRSs', rep(NA, 1))) 
tabletext = cbind(
  c("Category", "\n", forest.data$sub),
  c("Predictor", "\n", AFR_summary_tbl$mark),
  c("AUCPR (95% CI)", "\n", ci_text)
)
pop.cols = c("black","black","black")
```

```{r}
pdf(paste0(output_path, "AFR_AUCPR_Forest.pdf"), width = 8, height = 4.5)
forestplot(labeltext = tabletext, graph.pos = 1,
           mean = c(NA, NA, forest.data$mean_aucpr), 
           lower = c(NA, NA, forest.data$lower_CI),
           upper = c(NA, NA, forest.data$upper_CI),
           
           xticks = c(0.20, 0.30, 0.40, 0.50, 0.60),
           zero = max(forest.data$mean_aucpr),
           title = "Figure 2. Forest plot of models' AUCPR for AFR",
           xlab = "AUPRC with 95% CIs from bootstrapping",
           
           txt_gp = fpTxtGp(label = list(gpar(fontface = "bold", cex = 0.8, 
                                              fontfamily = "serif"),
                                         gpar(cex = 0.8, fontfamily = "serif"),
                                         gpar(cex = 0.8, fontfamily = "serif")),
                            ticks = gpar(cex = 0.6, fontfamily = "serif"),
                            xlab = gpar(cex = 0.7, fontfamily = "serif"),
                            title = gpar(cex = 1, fontfamily = "serif")),
           col = fpColors(text = pop.cols[c(1, 1, forest.data$class)], 
                          box  ="black",
                          lines = "black", 
                          zero ="gray50"),
           
           cex = 0.2, lineheight = "auto", boxsize = 0.25, 
           lwd.ci = 1, ci.vertices = TRUE, ci.vertices.height = 0.15)
dev.off()
```

### 2) AUC
```{r}
prs_summary_tbl = PRS_bootstrap_filter %>% 
  select(AUC, mark, bootstrap) %>% group_by(mark) %>% 
  summarise(mean_auc = mean(AUC), 
            lower_CI = quantile(AUC, 0.025),
            upper_CI = quantile(AUC, 0.975)) 
snp_summary_tbl = SNP_bootstrap_full %>% 
  select(AUC, mark, bootstrap) %>% group_by(mark) %>% 
  summarise(mean_auc = mean(AUC), 
            lower_CI = quantile(AUC, 0.025),
            upper_CI = quantile(AUC, 0.975)) 
AFR_summary_tbl = rbind(prs_summary_tbl, snp_summary_tbl) %>% 
  filter(mark %in% desired_order) %>% 
  arrange(factor(mark, levels = desired_order))
# Add CI to the table
ci_text = paste0(round(AFR_summary_tbl$mean_auc, 3), " (", 
                 round(AFR_summary_tbl$lower_CI, 3), ", ", 
                 round(AFR_summary_tbl$upper_CI, 3), ")")
```

```{r}
# build a forest plot 
forest.data = AFR_summary_tbl %>% 
  dplyr::select(mean_auc, lower_CI, upper_CI) %>% 
  mutate(sub = c("e4count",
    'EUR single PRS', rep(NA, 1), 
    'AFR single PRS' , rep(NA, 1),
    'Transethnic single PRS', rep(NA, 1),
    'AD combined PRSs', rep(NA, 1),
    'Neuro combined PRSs', rep(NA, 1),
    'AD SNPs', rep(NA, 1),
    'Neuro PRSs', rep(NA, 1))) 
tabletext = cbind(
  c("Category", "\n", forest.data$sub),
  c("Predictor", "\n", AFR_summary_tbl$mark),
  c("AUC (95% CI)", "\n", ci_text)
)
pop.cols = c("black","black","black")
```

```{r}
pdf(paste0(output_path, "AFR_AUC_Forest.pdf"), width = 8, height = 4.5)
forestplot(labeltext = tabletext, graph.pos = 1,
           mean = c(NA, NA, forest.data$mean_auc), 
           lower = c(NA, NA, forest.data$lower_CI),
           upper = c(NA, NA, forest.data$upper_CI),
           
           xticks = c(0.4, 0.5, 0.6, 0.7, 0.8),
           zero = max(forest.data$mean_auc),
           title = "Figure 2. Forest plot of models' AUC for AFR",
           xlab = "AUC with 95% CIs from bootstrapping",
           
           txt_gp = fpTxtGp(label = list(gpar(fontface = "bold", cex = 0.8, 
                                              fontfamily = "serif"),
                                         gpar(cex = 0.8, fontfamily = "serif"),
                                         gpar(cex = 0.8, fontfamily = "serif")),
                            ticks = gpar(cex = 0.6, fontfamily = "serif"),
                            xlab = gpar(cex = 0.7, fontfamily = "serif"),
                            title = gpar(cex = 1, fontfamily = "serif")),
           col = fpColors(text = pop.cols[c(1, 1, forest.data$class)], 
                          box  ="black",
                          lines = "black", 
                          zero ="gray50"),
           
           cex = 0.2, lineheight = "auto", boxsize = 0.25, 
           lwd.ci = 1, ci.vertices = TRUE, ci.vertices.height = 0.15)
dev.off()
```

## 3. P-value calculation
```{r}
PRS_bootstrap_update = PRS_bootstrap_filter %>% select(-bootstrap) 
PRS_bootstrap_update$bootstrap = rep(seq(1, 1000), each = 21)

SNP_bootstrap_update = SNP_bootstrap_full %>% select(-bootstrap)
SNP_bootstrap_update$bootstrap = rep(seq(1, 1000), each = 8)

combine_summary = rbind(PRS_bootstrap_update, SNP_bootstrap_update) 
dim(combine_summary) # dim = (29000, 30)

prs_columns = c("PRS.EUR.indsig", "PRS.EUR.map", "PRS.AFR.indsig", 
                "PRS.AFR.map", "PRS.trans.indsig", "PRS.trans.map",
                "PRS.AD.indsig.combine", "PRS.AD.map.combine",
                "PRS.neuro.indsig.combine", "PRS.neuro.map.combine")
snp_columns = c("SNP.AD.indsig", "SNP.AD.map", "SNP.all.indsig",
                "SNP.all.map")
aucpr_p = combine_summary %>%  
  select(mark, bootstrap, AUCPR) %>%
  pivot_wider(names_from = mark, values_from = AUCPR)
aucpr_p$max_prs = pmap_dbl(aucpr_p[prs_columns], max)
aucpr_p$max_snp = pmap_dbl(aucpr_p[snp_columns], max)
aucpr_p_summary = aucpr_p %>% 
  select(bootstrap, e4count, max_prs, max_snp, SNP.all.map) %>%
  mutate(prs_p = if_else(max_prs < SNP.all.map, 1, 0),
         e4count_p = if_else(e4count < SNP.all.map, 1, 0),
         map_is_max = if_else(max_snp <= SNP.all.map & max_prs < SNP.all.map, 1, 0)) 

table(aucpr_p_summary$e4count_p)[1]/1000                          
table(aucpr_p_summary$prs_p)[1]/1000
table(aucpr_p_summary$map_is_max)[1]/1000
```

```{r}
auc_p = combine_summary %>%  
  select(mark, bootstrap, AUC) %>%
  pivot_wider(names_from = mark, values_from = AUC)
auc_p$max_prs = pmap_dbl(auc_p[prs_columns], max)
auc_p$max_snp = pmap_dbl(auc_p[snp_columns], max)
auc_p_summary = auc_p %>% 
  select(bootstrap, e4count, max_prs, max_snp, SNP.all.map) %>%
  mutate(prs_p = if_else(max_prs < SNP.all.map, 1, 0),
         e4count_p = if_else(e4count < SNP.all.map, 1, 0),
         map_is_max = if_else(max_snp <= SNP.all.map & max_prs < SNP.all.map, 1, 0)) 

table(auc_p_summary$e4count_p)[1]/1000                          
table(auc_p_summary$prs_p)[1]/1000
table(auc_p_summary$map_is_max)[1]/1000
```


## 4. Selected SNPs
We ran bootstrapping (x1000) to get SNP map model (feature selected with 95%
confidence).
```{r}
# source(paste0(code_path, "snp_all_map_selection.R"))
load(file = paste0(raw_data_path, "modeling/AFR/feature_map_full.rda"))
max_boot = max(feature_map_full$bootstrap)
snps_map_summary = feature_map_full %>% 
  group_by(variable) %>% 
  summarise(n = n(), mean_imp = mean(percentage, na.rm = T), 
            lower_imp = quantile(percentage, 0.025, na.rm = T),
            upper_imp = quantile(percentage, 0.975, na.rm = T)) %>%
  arrange(desc(mean_imp)) %>% filter(n >= max_boot * 0.95) %>% 
  mutate(var_imp = paste0(round(mean_imp, 3), " (", 
                 round(lower_imp, 3), ", ", round(upper_imp, 3), ")")) %>% 
  select(variable, var_imp) %>% filter(variable != "Intercept")
dim(snps_map_summary) # dim = (10,2)
```

```{r}
load(file = paste0(raw_data_path, "FUMA_outputs/overlap_fuma.rda"))
load(file = paste0(raw_data_path, "Interpretation/info_tbl/snp_info_full.rda"))
AFR_SNP_map = overlap_fuma %>% 
  filter(chr_pos_name %in% snps_map_summary$variable) %>% 
  select(chr_pos_name, rsID, chr_num, pos, phenotype) %>% 
  left_join(snp_info_full, by = c("rsID" = "rsID")) %>% unique() %>% 
  left_join(snps_map_summary, by = c("chr_pos_name" = "variable")) %>% 
  arrange(factor(chr_pos_name, levels = snps_map_summary$variable)) %>% 
  select(var_imp, rsID, chr_num, pos, CADD, nearestGene, dist, func, posMapFilt, 
         eqtlMapFilt, ciMapFilt, phenotype) %>% mutate(mark = "x") %>% 
  pivot_wider(names_from = phenotype, values_from = mark) 
dim(AFR_SNP_map) # dim = (10,17)
write.table(AFR_SNP_map, 
            file = paste0(output_path, "interpretation/AFR_SNP_map.txt"),
            sep = "\t", quote = F, row.names = F, col.names = T)
```
